### 2024-11-12 #7455 @arfon
Copying here a note from the submission form:

> We are submitting PowNet 2.0, a power system modeling framework. The previous PowNet was published in the Open Research Journal (https://openresearchsoftware.metajnl.com/articles/10.5334/jors.302). However, this new version has undergone *a complete rewrite* with significant changes in design, functionality, and underlying mathematical formulations. PowNet 2.0 shares almost no similarity to its predecessor.
