### 2024-11-13 #7455 @phumthep
Thank you @arfon for starting the reviewing process! Seeing the number of backlogs, I have also volunteered for reviewing. As for reviewers in the list, these people might be able to provide feedback our work (power system modeling, optimization)

- Isaac Bromley-Dulfano
- Josh Wong
- Camilo Chacón Sartori
- blthayer
- Vissarion Fisikopoulos
