### 2024-12-11 7e38a55
Remove main in favor of run_quickstart.py
