### 2024-12-26 5d2f6df
Fix recording the initial condition
The previous implementation incorrectly uses the simulation horizon for calculations. The correct parameter is using 24-hours since we do a rolling simulation to simulate day-ahead dispatch.
