+++ src/pownet/core/output.py
@@ -8,48 +8,22 @@ from pownet.data_utils import get_dates, get_fuel_mix_order
 
 
 class OutputProcessor:
-    def __init__(self, inputs: SystemInput) -> None:
+    def __init__(self) -> None:
         """Initialize the OutputProcessor object."""
-        self.inputs: SystemInput = inputs
-        self.year: int = inputs.year
-        self.fuelmap: dict = inputs.fuelmap
-        self.nodal_hourly_demand: pd.DataFrame = inputs.demand
+        self.year: int = None
+        self.fuelmap: dict = {}
+        self.dates: pd.DataFrame = pd.DataFrame()
 
-        # Sum across each month to get the monthly dispatch
-        # For processing dataframes
-        self.dates: pd.DataFrame = get_dates(year=self.year)
+    def load(self, inputs: SystemInput) -> None:
+        """Load the input data."""
+        self.fuelmap = inputs.fuelmap
+        self.year = inputs.year
+
+        self.dates = get_dates(year=self.year)
         self.dates.index += 1
 
-        self.node_variables: pd.DataFrame = None
-
-        self.total_dispatch: pd.DataFrame = None
-        self.monthly_dispatch: pd.DataFrame = None
-        self.daily_dispatch: pd.DataFrame = None
-
-        self.total_demand: pd.Series = None
-        self.monthly_demand: pd.Series = None
-        self.daily_demand: pd.Series = None
-
-        # Define the order of fuel mix. Baseload at the bottom,
-        # renewables in the middle, then peaker plants, and shortfall
-        self.fuel_mix_order: list[str] = None
-        self.unit_status: pd.DataFrame = None
-
-    def load_from_dataframe(
-        self,
-        node_var_df: pd.DataFrame,
-    ) -> None:
-        """
-        Process node-specific variables from PowNet.
-
-        Args:
-            node_var_df (pd.DataFrame): The dataframe containing the node-specific variables.
-
-        Returns:
-            None
-        """
-        # Extract the dispatch of generators (thermal units and renewables)
-        dispatch_vars = [
+    def _get_power_variables(self, node_variables: pd.DataFrame) -> pd.DataFrame:
+        power_vars = {
             "pthermal",
             "psolar",
             "pwind",
@@ -57,203 +31,389 @@ class OutputProcessor:
             "pimp",
             "pos_pmismatch",
             "neg_pmismatch",
-        ]
-        self.node_variables = node_var_df.loc[
-            node_var_df["vartype"].isin(dispatch_vars)
+            "pdischarge",
+            "pcharge",
+        }
+        power_variables = node_variables.loc[
+            node_variables["vartype"].isin(power_vars)
         ].reset_index(drop=True)
-        self.node_variables["fuel_type"] = self.node_variables.apply(
+        power_variables["fuel_type"] = power_variables.apply(
             lambda x: self.fuelmap.get(x["node"], None), axis=1
         )
-        # Assign import and slack fuel types as they are not in the fuelmap dictionary
-        self.node_variables.loc[
-            self.node_variables["vartype"] == "pimp", "fuel_type"
-        ] = "import"
-        self.node_variables.loc[
-            self.node_variables["vartype"] == "pos_pmismatch", "fuel_type"
-        ] = "shortfall"
-        self.node_variables.loc[
-            self.node_variables["vartype"] == "neg_pmismatch", "fuel_type"
-        ] = "curtailment"
+        # These variables are missing in SystemInputs.fuelmap
+        vartype_to_fuel_type = {
+            "pimp": "import",
+            "pos_pmismatch": "shortfall",
+            "neg_pmismatch": "curtailment",
+            "pdischarge": "discharging",
+            "pcharge": "charging",
+        }
+        for vartype, fuel_type in vartype_to_fuel_type.items():
[…]
