+++ src/pownet/modeling/objfunc.py
@@ -14,26 +14,31 @@ def get_thermal_fixed_coeff(
 ) -> dict:
     """Fixed cost is a function of rated capacity and fixed cost per unit"""
     return {
-        (unit_g, t): thermal_rated_capacity[unit_g] * thermal_fixed_cost[unit_g]
+        (unit, t): thermal_rated_capacity[unit] * thermal_fixed_cost[unit]
         for t in timesteps
-        for unit_g in thermal_units
+        for unit in thermal_units
     }
 
 
 def get_thermal_opex_coeff(
+    step_k: int,
     timesteps: range,
     thermal_units: list,
     thermal_opex: dict,
-    thermal_fuel_cost: dict,
+    fuel_contracts: dict,
+    contract_costs: dict,
     thermal_heat_rate: dict,
 ) -> dict:
     """Variable cost is a function of fuel cost, heat rate, and opex"""
     # TODO: thermal_fuel_cost should be a time-dependent parameter
     return {
-        (unit_g, t): (thermal_fuel_cost[unit_g] * thermal_heat_rate[unit_g])
-        + thermal_opex[unit_g]
+        (unit, t): (
+            contract_costs[(fuel_contracts[unit], t + (step_k - 1) * 24)]
+            * thermal_heat_rate[unit]
+        )
+        + thermal_opex[unit]
         for t in timesteps
-        for unit_g in thermal_units
+        for unit in thermal_units
     }
 
 
@@ -45,18 +50,18 @@ def get_thermal_startup_coeff(
 ) -> dict:
     """Startup cost is a function of rated capacity and startup cost per unit"""
     return {
-        (unit_g, t): thermal_rated_capacity[unit_g] * thermal_startup_cost[unit_g]
+        (unit, t): thermal_rated_capacity[unit] * thermal_startup_cost[unit]
         for t in timesteps
-        for unit_g in thermal_units
+        for unit in thermal_units
     }
 
 
 def get_marginal_cost_coeff(
-    inputs: SystemInput,
     step_k: int,
     timesteps: range,
     units: list,
-    attribute: str,
+    nondispatch_contracts: dict,
+    contract_costs: dict,
 ) -> dict:
     """
     Generic helper function to calculate coefficients based on marginal cost or a similar attribute for a list of units
@@ -69,7 +74,7 @@ def get_marginal_cost_coeff(
         A dictionary mapping (unit, t) tuples to the calculated coefficients
     """
     return {
-        (unit, t): getattr(inputs, attribute).loc[t + (step_k - 1) * 24, unit]
+        (unit, t): contract_costs[(nondispatch_contracts[unit], t + (step_k - 1) * 24)]
         for t in timesteps
         for unit in units
     }
