+++ src/pownet/core/input.py
@@ -7,6 +7,7 @@ import os
 import textwrap
 
 import gurobipy as gp
+from gurobipy import GRB
 import pandas as pd
 
 logger = logging.getLogger(__name__)
@@ -22,10 +23,11 @@ class SystemInput:
         use_spin_var: bool = True,
         dc_opf: str = "kirchhoff",
         spin_reserve_factor: float = 0.15,
+        spin_reserve_mw: float = None,
         line_loss_factor: float = 0.075,
         line_capacity_factor: float = 0.9,
         load_shortfall_penalty_factor: float = 1000,
-        load_curtail_penalty_factor: float = 10,
+        load_curtail_penalty_factor: float = 900,
         spin_shortfall_penalty_factor: float = 1000,
     ) -> None:
         """This class reads the input data for the power system model."""
@@ -48,8 +50,11 @@ class SystemInput:
             )
         self.dc_opf: str = dc_opf
 
-        # The spin reserve factor is a fraction of total demand
+        # The spin reserve factor is a fraction of total demand, so should be between 0 and 1
+        if not 0 <= spin_reserve_factor <= 1:
+            raise ValueError("PowNet: Spin reserve factor must be between 0 and 1.")
         self.spin_reserve_factor: float = spin_reserve_factor
+        self.spin_reserve_mw: float = spin_reserve_mw
 
         # The line loss factor is also called the inefficiency factor.
         # It is the fraction of power lost in the transmission line.
@@ -73,7 +78,6 @@ class SystemInput:
 
         # Thermal units
         self.thermal_unit_node: dict[str, str] = {}
-        self.thermal_fuel_cost: dict[str, str] = {}
 
         self.thermal_fixed_cost: dict[str, float] = {}
         self.thermal_opex: dict[str, float] = {}
@@ -81,7 +85,7 @@ class SystemInput:
         self.thermal_heat_rate: dict[str, float] = {}
 
         self.thermal_rated_capacity: dict[str, float] = {}
-        self.thermal_derated_capacity: dict[str, float] = {}
+        self.thermal_derated_capacity: pd.DataFrame = pd.DataFrame()
         self.thermal_min_capacity: dict[str, float] = {}
 
         self.TD: dict[str, int] = {}
@@ -91,62 +95,114 @@ class SystemInput:
         self.RD: dict[str, float] = {}
         self.RU: dict[str, float] = {}
 
-        # Other data
+        # Energy storage
+        self.ess_unit_attach: dict[str, str] = {}
+        self.ess_attach_unit: dict[str, list[str]] = {}
+
+        self.ess_hydro_units: dict[str, list[str]] = {}
+        self.ess_daily_hydro_units: dict[str, list[str]] = {}
+        self.ess_solar_units: dict[str, list[str]] = {}
+        self.ess_wind_units: dict[str, list[str]] = {}
+        self.ess_thermal_units: dict[str, list[str]] = {}
+        self.ess_substation_units: dict[str, list[str]] = {}
+
+        self.ess_max_charge: dict[str, float] = {}
+        self.ess_max_discharge: dict[str, float] = {}
+        self.ess_max_capacity: dict[str, float] = {}
+        self.ess_min_capacity: dict[str, float] = {}
+        self.ess_charge_efficiency: dict[str, float] = {}
+        self.ess_discharge_efficiency: dict[str, float] = {}
+        self.ess_self_discharge_rate: dict[str, float] = {}
+        self.ess_derated_capacity: pd.DataFrame = pd.DataFrame()
+
+        self.ess_unit_node: dict[str, str] = {}
+
+        # Demand
         self.demand: pd.DataFrame = pd.DataFrame()
[…]

[…]
