+++ pyproject.toml
@@ -20,5 +20,7 @@ dependencies = [
     "scipy >= 1.11.3",
     "scikit-learn >= 1.5.2",
     "statsmodels >= 0.14.4",
+    "geopandas >= 1.0.1",
+    "contextily >= 1.6.2",
 ]
 requires-python = ">=3.10"
