### 2025-01-24 24ec512
Fix legend in visualizer.py
