+++ pyproject.toml
@@ -1,6 +1,12 @@
 [build-system]
-requires = ["setuptools>=61.0.0", "wheel"]
+requires = ["setuptools>=61.0.0", "setuptools-scm", "wheel"]
 build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[tool.setuptools.package-data]
+pownet = ["database/*.csv"]
 
 [project]
 name = "cisl-pownet"
