### 2025-01-26 149baa4
Add num_sim_days as input arg
This change allows more flexibility when creating input files.
