### 2025-01-26 b8aa8cd
Add num_sim_days as input arg
This change allows more flexibility when creating input files.
