### 2025-01-28 5313c9a
Prepare for uploading v2.2 to PyPI
