### 2025-02-19 8428d1d
Update example scripts
