### 2025-04-03 #7455 @arfon
@prakaa @blthayer @hellemo @ijbd :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is PowNet 2.0: A production cost modeling framework for large-scale power systems

Based on your experience, and your prior experience with JOSS, we think you might be able to provide a great review of this submission. Please let me know if you think you can help us out!

Many thanks
Arfon
