### 2025-04-10 #8035 @editorialbot
[REVIEW]: PowNet 2.0: A production cost modeling framework for large-scale power systems
**Submitting author:** @phumthep (<a href="http://orcid.org/0000-0001-6412-7017">Phumthep Bunnak</a>)
**Repository:** https://github.com/Critical-Infrastructure-Systems-Lab/PowNet
**Branch with paper.md** (empty if default branch): master
**Version:** v2.1
