### 2025-04-21 #8035 @prakaa
My comments on the paper:

1. In my opinion lines 8-12 in the summary (describing PCMs) could be condensed to improve readability and make the summary more concise. I think focusing on what PCMs do, as opposed to what they do not, would be helpful since dynamics and investment/planning models aren't described in the paper. A brief definition of "short-term" or the timeframes involved would add clarity too.
2. There appears to be no space between the name of the software/package and the in-text citation for some of the software/packages in the Statement of need.
3. In the Mathematical Formulations section of the Statement of need, I think it would make more sense for these references to be in parentheses. You are not referring directly to the cited texts.
4. Examining the needs in lines 35-39, 1) and 4) are important but arguably handled relatively well by the software and packages you have listed in the preceding paragraph. Would it better highlight the contribution of this package to focus on 2) (solver flexibility & solution heuristics), 3) (module to generate synthetic time series to use as inputs) and on the water-energy focus (reservoir simulation - anything more?)? The Modelling Framework subheading could be incorporated into the summary.
5. Following on from the point above, do you think the listed features are the main points of differentiation from Sienna, PyPSA and PLEXOS?
6. If the time series model in Figure 1 is linked to the ability to generate stochastic inputs as described in the Statement of need, I think it could be clearer in Figure 1 that this is the purpose of the "Time series model"
7. Figure 1 reference on line 70 should be fixed
