### 2025-04-21 #8035 @prakaa
My comments on documentation:

1. A brief mention of the target audience and/or comparison with similar packages could be handy for potential users
2. As a general comment, I think there could be more detail in the documentation. Some of this could be inferred from looking at the source code or the example input data for the dummy model, but I think it would be of general benefit to a user to have more detail. A non-exhaustive list of things that could be covered: format/requirements of input data (csvs/dataframes for costs, generators, nodes, lines, etc.), format of the output data and features mentioned in the JOSS paper (reservoir modelling, synthetic time series generation, the use of alternative formulations, integer rounding solution heuristic.
3. The example did not work for me by just copying the code. The if statement that checks whether the output path exists and then creates it needs indentation. The input and output folder paths did not work on my machine (Ubuntu 24), could you use `pathlib` instead? I changed these and changed the solver to `highs` since I am no longer an academic with access to Gurobi, but it appears that the model can't be built without a Gurobi licence. Is this right (see screenshot below)

![Image](https://github.com/user-attachments/assets/ac35e740-4dca-4636-8bae-a2479602e5c3)

5. Similar to the documentation, the code example requires some additional guidance/explanation. First off, the user needs to clone the source code repo to access the dummy model directory (if they have pip installed or installed with a dependency management tool). Secondly, it's unclear what is being run - a summary of the scenario would be ideal, but at the very least a link and brief description of the inputs would help.
6. In the API reference, some classes and methods are documented, but some are not. Is there some reason for this?
7. I couldn't find instructions for development - e.g. how to run tests.
8. I couldn't find contributing guidelines
