### 2025-04-22 df3ca69
Revise example scripts
