+++ scripts/run_pownet.py
@@ -1,3 +1,4 @@
+import os
 from datetime import datetime
 from pownet.core import (
     DataProcessor,
@@ -15,16 +16,18 @@ sim_horizon = 24
 model_name = "dummy"
 steps_to_run = 2  # Default is None
 do_plot = True
+
+input_folder = os.path.join("..", "model_library")
 #######################
 
 if to_process_inputs:
     data_processor = DataProcessor(
-        input_folder="../model_library", model_name=model_name, year=2016, frequency=50
+        input_folder=input_folder, model_name=model_name, year=2016, frequency=50
     )
     data_processor.execute_data_pipeline()
 
 inputs = SystemInput(
-    input_folder="../model_library",
+    input_folder=input_folder,
     model_name=model_name,
     year=2016,
     sim_horizon=sim_horizon,
@@ -40,8 +43,6 @@ model_builder = ModelBuilder(inputs)
 record = SystemRecord(inputs)
 
 build_times = []
-opt_times = []
-objvals = []
 init_conditions = create_init_condition(inputs.thermal_units)
 
 if steps_to_run is None:
@@ -60,41 +61,35 @@ for step_k in range(1, steps_to_run):
             init_conds=init_conditions,
         )
     build_times.append((datetime.now() - start_time).total_seconds())
-
     power_system_model.optimize(mipgap=0.001)
 
     # Raise an error if the model is not feasible
     if not power_system_model.check_feasible():
         raise ValueError("Model is not feasible.")
 
-    objvals.append(power_system_model.get_objval())
-    opt_times.append(power_system_model.get_runtime())
-
     record.keep(
         runtime=power_system_model.get_runtime(),
         objval=power_system_model.get_objval(),
         solution=power_system_model.get_solution(),
-        lmp=power_system_model.solve_for_lmp(),
         step_k=step_k,
     )
     init_conditions = record.get_init_conds()
 
 
 # Process the results
+node_variables = record.get_node_variables()
+
 output_processor = OutputProcessor()
 output_processor.load(inputs)
-node_variables = record.get_node_variables()
+hourly_generation = output_processor.get_hourly_generation(node_variables)
+hourly_demand = output_processor.get_hourly_demand(inputs.demand)
+
+# Similar methods are available to get monthly values
+# monthly_generation = output_processor.get_monthly_generation(node_variables)
+# monthly_demand = output_processor.get_monthly_demand(inputs.demand)
 
 # Visualize the results
 if do_plot:
     visualizer = Visualizer(inputs.model_id)
-    if steps_to_run <= 3:
-        visualizer.plot_fuelmix_bar(
-            dispatch=output_processor.get_hourly_generation(node_variables),
-            demand=output_processor.get_hourly_demand(inputs.demand),
-        )
-    else:
-        visualizer.plot_fuelmix_area(
-            dispatch=output_processor.get_monthly_generation(node_variables),
-            demand=output_processor.get_monthly_demand(inputs.demand),
-        )
+
+    visualizer.plot_fuelmix_bar(dispatch=hourly_generation, demand=hourly_demand)
