+++ docs/requirements.txt
@@ -1,4 +1,3 @@
 sphinx>=4.3.0
 sphinx-rtd-theme>=0.5.1
-sphinx-toolbox>=v0.8.0
-
+sphinx-toolbox>=0.8.0
