+++ pyproject.toml
@@ -21,7 +21,7 @@ dependencies = [
     "highspy >= 1.7.2",
     "matplotlib >= 3.7.1",
     "networkx >= 3.1.0",
-    "numpy == 2.2.0",
+    "numpy >= 2.2.0",
     "pandas >= 2.1.1",
     "pmdarima >= 2.0.4",
     "scipy >= 1.11.3",
