+++ pyproject.toml
@@ -23,7 +23,6 @@ dependencies = [
     "networkx >= 3.1.0",
     "numpy >= 2.2.0",
     "pandas >= 2.1.1",
-    "pmdarima >= 2.0.4",
     "scipy >= 1.11.3",
     "scikit-learn >= 1.5.2",
     "statsmodels >= 0.14.4",
