+++ .github/workflows/python-app.yml
@@ -23,7 +23,6 @@ jobs:
       uses: actions/setup-python@v3
       with:
         python-version: "3.10"
-        cache: 'false'
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
