+++ src/pownet/core/builder.py
@@ -2,17 +2,13 @@
 """
 
 from .input import SystemInput
-import logging
 
 from gurobipy import GRB
 import gurobipy as gp
 import pandas as pd
 
 import pownet.modeling as modeling
-from pownet.modeling import PowerSystemModel
 from pownet.data_utils import get_unit_hour_from_varnam, get_edge_hour_from_varname
-
-logger = logging.getLogger(__name__)
 
 
 class ModelBuilder:
@@ -35,7 +31,7 @@ class ModelBuilder:
     def __init__(self, inputs: SystemInput) -> None:
         self.inputs = inputs
         self.timesteps = range(1, self.inputs.sim_horizon + 1)
-        self.model: gp.Model = None
+        self.model: gp.Model = gp.Model(self.inputs.model_id)
 
         # Variables
         # Thermal units
@@ -51,6 +47,7 @@ class ModelBuilder:
         # Energy storage
         self.pcharge = gp.tupledict()
         self.pdischarge = gp.tupledict()
+        self.pdischarge_shortfall = gp.tupledict()
 
         self.charge_state = gp.tupledict()  # State of charge
         self.ucharge = gp.tupledict()  # Charging indicator
@@ -92,6 +89,7 @@ class ModelBuilder:
 
         self.load_shortfall_penalty_expr = gp.LinExpr()
         self.spin_shortfall_penalty_expr = gp.LinExpr()
+        self.pdischarge_shortfall_penalty_expr = gp.LinExpr()
 
         # Thermal unit constraints
         self.c_link_uvw_init = gp.tupledict()
@@ -161,6 +159,7 @@ class ModelBuilder:
         -----------------
         - `pcharge`: Power charging an energy storage system. Unit: MW.
         - `pdischarge`: Power discharging an energy storage system. Unit: MW.
+        - `pdischarge_shortfall`: Shortfall in discharging power. Unit: MW.
         - `charge_state`: State of charge of an energy storage system. Unit: MWh.
         - `ucharge`: Indicator that an ESS is charging. Unitless.
         - `udischarge`: Indicator that an ESS is discharging. Unitless.
@@ -205,28 +204,33 @@ class ModelBuilder:
             (
                 "phydro_curtail",
                 self.inputs.hydro_units,
-                self.inputs.hydro_contracted_capacity,
+                self.inputs.hydro_max_capacity,
             ),
             ("psolar", self.inputs.solar_units, self.inputs.solar_contracted_capacity),
             (
                 "psolar_curtail",
                 self.inputs.solar_units,
-                self.inputs.solar_contracted_capacity,
+                self.inputs.solar_max_capacity,
             ),
             ("pwind", self.inputs.wind_units, self.inputs.wind_contracted_capacity),
             (
                 "pwind_curtail",
                 self.inputs.wind_units,
-                self.inputs.wind_contracted_capacity,
+                self.inputs.wind_max_capacity,
             ),
             ("pimp", self.inputs.import_units, self.inputs.import_contracted_capacity),
             (
                 "pimp_curtail",
                 self.inputs.import_units,
-                self.inputs.import_contracted_capacity,
+                self.inputs.import_max_capacity,
             ),
             ("pcharge", self.inputs.storage_units, self.inputs.ess_max_charge),
             ("pdischarge", self.inputs.storage_units, self.inputs.ess_max_discharge),
+            (
+                "pdischarge_shortfall",
+                self.inputs.storage_units,
+                self.inputs.ess_max_discharge,
+            ),
         ]
 
         for varname, units, capacity_dict in var_with_u_tuples:
@@ -474,6 +478,12 @@ class ModelBuilder:
             self.inputs.spin_shortfall_penalty_factor * gp.quicksum(self.spin_shortfall)
         )
 
+        # Penalize discharge shortfall
[…]

[…]
