+++ docs/source/conf.py
@@ -1,5 +1,7 @@
 import sys
 import os
+from importlib.metadata import version as get_version
+
 
 sys.path.insert(0, os.path.abspath("../../src/"))
 
@@ -23,57 +25,46 @@ autodoc_mock_imports = [
 # -- Project information
 
 project = "PowNet"
-copyright = "2024, Critical Infrastructure Systems (CIS) Lab, Cornell University"
+copyright = "2021-2025, Critical Infrastructure Systems (CIS) Lab, Cornell University"
 author = "Critical Infrastructure Systems Lab (CIS), Cornell University"
 
-release = "2.0"
-version = "2.0.0"
+# TODO: Show version and release in the documentation
+release = get_version(project)
+version = ".".join(release.split(".")[:1])
 
 # -- General configuration
 
-github_username = "HishamEldardiry"
-github_repository = "PowNet_Documentation"
-
-
 extensions = [
-    "sphinx.ext.duration",
-    "sphinx.ext.doctest",
     "sphinx.ext.autodoc",
-    "sphinx.ext.napoleon",
-    "sphinx_autodoc_typehints",
     "sphinx.ext.autosummary",
+    "sphinx.ext.doctest",
+    "sphinx.ext.duration",
+    "sphinx.ext.extlinks",
     "sphinx.ext.intersphinx",
     "sphinx.ext.mathjax",
+    "sphinx.ext.napoleon",
     "sphinx.ext.viewcode",
     "sphinx_rtd_theme",
-    "sphinx.ext.viewcode",
-    "sphinx_toolbox",
+    "nbsphinx",
+    "nbsphinx_link",
+    "sphinx_autodoc_typehints",
+    "sphinxcontrib.bibtex",
 ]
-
-intersphinx_mapping = {
-    "python": ("https://docs.python.org/3/", None),
-    "sphinx": ("https://www.sphinx-doc.org/en/master/", None),
-}
-intersphinx_disabled_domains = ["std"]
 
 templates_path = ["_templates"]
 html_static_path = ["_static"]
 
+# References are found here
+bibtex_bibfiles = ["references.bib"]
+
+# Create the page even when there is an error in the notebook
+nbsphinx_allow_errors = True
+
 # -- Options for HTML output
-
 html_theme = "sphinx_rtd_theme"
-html_theme_options = {"version": "", "release": "", "navigation_depth": 2}
+html_theme_options = {}
 
-html_context = {
-    "display_github": True,  # Add 'Edit on Github' link instead of 'View page source'
-    "last_updated": False,
-    "commit": False,
-}
-
-# -- html_show_sphinx = False
-
-# -- Options for EPUB output
-epub_show_urls = "footnote"
+html_show_sphinx = False
 
 
 # -- customize the CSS styling
