+++ docs/source/technical/time_series_models.rst
@@ -1,0 +1,8 @@
+
+.. autosummary::
+    :toctree: _source/
+    
+**Time series modeling**
+=========================
+
+Loren ipsum
