+++ docs/source/technical/dispatchable.rst
@@ -1,0 +1,8 @@
+
+.. autosummary::
+    :toctree: _source/
+    
+**Dispatchable/Non-dispatchable generators**
+===============================================
+
+Loren ipsum
