+++ docs/requirements.txt
@@ -1,3 +1,5 @@
-sphinx>=4.3.0
-sphinx-rtd-theme>=0.5.1
-sphinx-toolbox>=0.8.0
+sphinx >= 8.1.0
+sphinx-rtd-theme >= 3.0.2
+nbsphinx >= 0.9.7
+nbsphinx_link >= 1.3.1
+sphinxcontrib-bibtex >= 2.6.3

[…]
