+++ docs/requirements.txt
@@ -3,3 +3,4 @@ sphinx-rtd-theme >= 3.0.2
 nbsphinx >= 0.9.7
 nbsphinx_link >= 1.3.1
 sphinxcontrib-bibtex >= 2.6.3
+sphinx_autodoc_typehints
