+++ docs/requirements.txt
@@ -4,3 +4,4 @@ nbsphinx >= 0.9.7
 nbsphinx_link >= 1.3.1
 sphinxcontrib-bibtex >= 2.6.3
 sphinx_autodoc_typehints
+ipykernel
