### 2025-04-29 31b3678
Add synthetic time series examples
Electricity demand and solar time series
