### 2025-05-01 Critical-Infrastructure-Systems-Lab/PowNet#20 @ijbd
Broken Links in README
https://github.com/openjournals/joss-reviews/issues/8035

I noticed the last three links in the readme seem to be broken:

![Image](https://github.com/user-attachments/assets/09688082-97d3-4fb5-a87b-b5e068eff99d)
