### 2025-05-01 Critical-Infrastructure-Systems-Lab/PowNet#21 @ijbd
Missing community guidelines
https://github.com/openjournals/joss-reviews/issues/8035

The readthedocs update is looking great! Seems like still in progress, but just based on the skeleton, wondering if you will still need a community guidelines/contributing page?
