### 2025-05-01 Critical-Infrastructure-Systems-Lab/PowNet#22 @ijbd
Paper Feedback
https://github.com/openjournals/joss-reviews/issues/8035

Here is my feedback for the paper:

**Paper Review**

Statement of Need
- Seconding @prakaa's comments about which features to (re)prioritize, I think you have a great structure for comparing your software with the existing solutions. In thinking through your edits, you might consider identifying why [your] particular and future research lead(s) to pownet 2.0 filling gaps left by other solutions. This is partially done already by some references in your feature descriptions, but just a note as you edit.
- Line 36 and 40 "...the ability to efficiently handle large-scale systems with numerous nodes and edges..." and "...specifically designed to run multiple simulations of large-scale power systems..."
    - These are vague functional claims, the second more so than the first. If these are limitiations of existing packages, which of your features responds to these needs? Again, why did/does your research lead you to addressing these?
- Line 50 "Selecting a set of computationally efficient mathematical formulations is an on-going research Chen et al. (2022)"
    - Change to "research agenda Chen et al. (2022)"?
- Line 63 "Reservoir simulation"
    - This is a very interesting feature, and all of the relevant studies and publications in the readthedocs appear to explore hydro and energy-water system coupling. Can you elaborate?

Software Design, Implementation, and Validation
- I don't see a discussion of validation, but former research publications and on-going research count! This could be a good place to list out those projects.
- Line 67 "The software is specifically designed to support power-water systems coupling"
    - This is another indicator to me that power-water systems coupling should have more emphasis in the statement of need and possibly summary.
