### 2025-05-09 4cb4fcc
feat(weekly_hydro): weekly hydro introduced
