+++ src/pownet/core/input.py
@@ -106,6 +106,7 @@ class SystemInput:
 
         self.ess_hydro_units: dict[str, list[str]] = {}
         self.ess_daily_hydro_units: dict[str, list[str]] = {}
+        self.ess_weekly_hydro_units: dict[str, list[str]] = {}
         self.ess_solar_units: dict[str, list[str]] = {}
         self.ess_wind_units: dict[str, list[str]] = {}
         self.ess_thermal_units: dict[str, list[str]] = {}
@@ -130,9 +131,12 @@ class SystemInput:
         # Hydropower (hourly and daily timeseries)
         self.hydro_contracted_capacity: dict[str, float] = {}
         self.hydro_capacity: pd.DataFrame = pd.DataFrame()
+        self.hydro_capacity_min: pd.DataFrame = pd.DataFrame()
         self.hydro_unit_node: dict[str, str] = {}
         self.daily_hydro_capacity: pd.DataFrame = pd.DataFrame()
         self.daily_hydro_unit_node: dict[str, str] = {}
+        self.weekly_hydro_capacity: pd.DataFrame = pd.DataFrame()
+        self.weekly_hydro_unit_node: dict[str, str] = {}
 
         self.solar_contracted_capacity: dict[str, float] = {}
         self.solar_capacity: pd.DataFrame = pd.DataFrame()
@@ -179,6 +183,7 @@ class SystemInput:
 
         self.hydro_must_take_units: list[str] = []
         self.daily_hydro_must_take_units: list[str] = []
+        self.weekly_hydro_must_take_units: list[str] = []
         self.solar_must_take_units: list[str] = []
         self.wind_must_take_units: list[str] = []
         self.import_must_take_units: list[str] = []
@@ -330,6 +335,9 @@ class SystemInput:
             elif attached in self.daily_hydro_unit_node.keys():
                 self.ess_daily_hydro_units[attached] = self.ess_attach_unit[attached]
 
+            elif attached in self.weekly_hydro_unit_node.keys():
+                self.ess_weekly_hydro_units[attached] = self.ess_attach_unit[attached]
+
             elif attached in self.solar_unit_node.keys():
                 self.ess_solar_units[attached] = self.ess_attach_unit[attached]
 
@@ -372,6 +380,7 @@ class SystemInput:
             self._add_units_to_node(node, self.thermal_unit_node)
             self._add_units_to_node(node, self.hydro_unit_node)
             self._add_units_to_node(node, self.daily_hydro_unit_node)
+            self._add_units_to_node(node, self.weekly_hydro_unit_node)
             self._add_units_to_node(node, self.solar_unit_node)
             self._add_units_to_node(node, self.wind_unit_node)
             self._add_units_to_node(node, self.import_unit_node)
@@ -468,6 +477,17 @@ class SystemInput:
                 )
             )
 
+        # Units with weekly timeseries
+        if os.path.exists(os.path.join(self.model_dir, "hydropower_weekly.csv")):
+            self.weekly_hydro_capacity, self.weekly_hydro_unit_node = (
+                self._load_capacity_and_update_fuelmap_and_get_unit_node(
+                    "hydropower_weekly.csv", fuel_type="hydropower"
+                )
+            )
+
+        if os.path.exists(os.path.join(self.model_dir, "hydro_capacity_min.csv")):
+            self.hydro_capacity_min = pd.read_csv(os.path.join(self.model_dir, "hydro_capacity_min.csv"))
+
         # Check that the names do not repeat across different types
         repeated_units = set(self.hydro_unit_node.keys()).intersection(
             self.daily_hydro_unit_node.keys()
@@ -475,6 +495,24 @@ class SystemInput:
         if repeated_units:
             raise ValueError(
[…]
