### 2025-05-09 54d7f92
Add estimate hydropower example
