+++ src/pownet/data_model/__init__.py
@@ -1,0 +1,7 @@
+"""This is the core module."""
+
+from .reservoir import ReservoirParams
+
+__all__ = [
+    "ReservoirParams",
+]
