+++ src/pownet/reservoir/reservoir.py
@@ -1,5 +1,4 @@
-""" reservoir.py: This module contains classes to simulate the operation of a reservoir.
-"""
+"""reservoir.py: Reservoir class"""
 
 import math
 import os
@@ -8,14 +7,13 @@ import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 
-from pownet.data_utils import remove_29feb
-from pownet.folder_utils import get_model_dir, get_reservoir_file
+from ..data_model import ReservoirParams
+
 from .solve_release import (
     solve_release_from_target_storage,
     solve_release_from_dispatch,
 )
 from .reservoir_functions import (
-    calc_min_flow,
     calc_target_level,
     calc_target_storage,
     calc_level_from_storage,
@@ -28,182 +26,123 @@ from .reservoir_functions import (
 
 
 class Reservoir:
-    """This class simulates a reservoir."""
+    """
+    Simulates the operation of a single reservoir, initialized using ReservoirParams.
+    The simulation horizon is set to 365 days with a daily time step.
+    """
 
-    def __init__(
-        self,
-        model_name: str,
-        reservoir_name: str,
-        upstream_flow: pd.Series = None,
-        sim_horizon: int = 365,
-    ) -> None:
+    def __init__(self, params: ReservoirParams) -> None:
+        """
+        Initialize the Reservoir object.
 
-        # Number of days in a year
-        self.sim_horizon = sim_horizon
+        Args:
+            params: A dataclass instance containing all static parameters
+                              and initial inflow timeseries for the reservoir.
+        """
+        self.sim_days = 365  # days in a year
+        self._time_index = pd.RangeIndex(start=1, stop=self.sim_days + 1, step=1)
 
-        self.model_name: str = model_name
-        self.name: str = reservoir_name
+        # --- Unpack parameters from the dataclass ---
+        self.name: str = params.name
+        self.min_day: int = params.min_day
+        self.max_day: int = params.max_day
+        self.min_level: float = params.min_level
+        self.max_level: float = params.max_level
+        self.max_head: float = params.max_head
+        self.max_storage: float = params.max_storage
+        self.max_release: float = params.max_release
+        self.max_generation: float = params.max_generation
+        self.turbine_factor: float = params.turbine_factor
 
-        # Upstream flow
-        self.upstream_flow: pd.Series = None
-        if upstream_flow is not None:
-            self.set_upstream_flow(upstream_flow)
-        else:
-            self.upstream_flow: pd.Series = pd.Series(
-                0, index=range(1, self.sim_horizon + 1)
-            )
+        self.upstream_units: list[str] = params.upstream_units
+        self.downstream_flow_fracs: dict[str, float] = params.downstream_flow_fracs
 
-        ############################
-        # Loaded parameters from CSV
-        ############################
+        # --- Set timeseries data ---
+        self.inflow_ts: pd.Series = params.inflow_ts
+        self.minflow_ts: pd.Series = params.minflow_ts
 
-        self.min_day: int = None
-        self.max_day: int = None
-        self.min_level: float = None  # in meters
-        self.max_level: float = None  # in meters
+        # --- Initialize upstream flow (can be set later) ---
+        self.outflow: pd.Series = pd.Series(0.0, index=self._time_index, dtype=float)
 
-        # Max head is the maximum distance between the water level and turbine
-        self.max_head: float = None  # in meters
-        self.max_storage: float = None  # in m3
-        self.max_release: float = None  # in m3/day
-        # Max generation is power in MW
-        self.max_generation: float = None
-        # Index of the reservoir's cascade level. Zero is the furthest upstream.
-        self.cascade_level: int = None
-        # Turbine efficiency
[…]
