+++ src/pownet/reservoir/__init__.py
@@ -1,5 +1,4 @@
-""" This is the reservoir module.
-"""
+"""This is the reservoir module."""
 
 from .reservoir import Reservoir
 from .basin import Basin
@@ -8,3 +7,4 @@ from .solve_release import (
     solve_release_from_target_storage,
     solve_release_from_dispatch,
 )
+from .manager import ReservoirManager
