### 2025-05-09 a8acd8b
Revise solve_release.py
