+++ src/pownet/core/simulation.py
@@ -6,7 +6,7 @@ from pownet.data_utils import (
 )
 from .builder import ModelBuilder
 from .data_processor import DataProcessor
-from .input import SystemInput
+from ..input import SystemInput
 from .output import OutputProcessor
 from .record import SystemRecord
 from .visualizer import Visualizer
@@ -261,62 +261,62 @@ class Simulator:
             output_folder=output_folder,
         )
 
-    # def reoperate(
-    #     self,
-    #     k: int,
-    #     builder: ModelBuilder,
-    #     init_conds: dict,
-    #     mip_gap: float = None,
-    #     timelimit: float = None,
-    # ):
-    #     raise NotImplementedError("Reoperation is not implemented yet.")
-    #     reop_converge = False
-    #     reop_k = 0
-    #     while not reop_converge:
-    #         print(f"\nReservoirs reoperation iteration {reop_k}")
-    #         print("New Capacity vs. Current Dispatch")
+    def reoperate(
+        self,
+        k: int,
+        builder: ModelBuilder,
+        init_conds: dict,
+        mip_gap: float = None,
+        timelimit: float = None,
+    ):
+        raise NotImplementedError("Reoperation is not implemented yet.")
+        reop_converge = False
+        reop_k = 0
+        while not reop_converge:
+            print(f"\nReservoirs reoperation iteration {reop_k}")
+            print("New Capacity vs. Current Dispatch")
 
-    #         # PowNet returns the hydropower dispatch in hourly resolution across the simulation horizon
-    #         hydro_dispatch, start_day, end_day = get_hydro_from_model(
-    #             self.model.model, k
-    #         )
-    #         # Convert to daily dispatch
-    #         hydro_dispatch = convert_to_daily_hydro(hydro_dispatch, start_day, end_day)
-    #         new_hydro_capacity = self.reservoir_operator.reoperate_basins(
-    #             pownet_dispatch=hydro_dispatch
-    #         )
+            # PowNet returns the hydropower dispatch in hourly resolution across the simulation horizon
+            hydro_dispatch, start_day, end_day = get_hydro_from_model(
+                self.model.model, k
+            )
+            # Convert to daily dispatch
+            hydro_dispatch = convert_to_daily_hydro(hydro_dispatch, start_day, end_day)
+            new_hydro_capacity = self.reservoir_operator.reoperate_basins(
+                pownet_dispatch=hydro_dispatch
+            )
 
-    #         for res in new_hydro_capacity.columns:
-    #             print(
-    #                 f"{res}: {round(new_hydro_capacity[res].sum(),2)} vs {round(hydro_dispatch[res].sum(),2)}",
-    #             )
+            for res in new_hydro_capacity.columns:
+                print(
+                    f"{res}: {round(new_hydro_capacity[res].sum(),2)} vs {round(hydro_dispatch[res].sum(),2)}",
+                )
 
-    #         max_deviation = (new_hydro_capacity - hydro_dispatch).abs().max()
-    #         # The tolerance for convergence should be 5% of the largest hydro capacity
-    #         reop_tol = 0.05 * new_hydro_capacity.max()
-    #         if (max_deviation <= reop_tol[max_deviation.index]).all():
-    #             reop_converge = True
-    #             print(f"PowNet: Day {k+1} - Reservoirs converged at iteration {reop_k}")
+            max_deviation = (new_hydro_capacity - hydro_dispatch).abs().max()
+            # The tolerance for convergence should be 5% of the largest hydro capacity
+            reop_tol = 0.05 * new_hydro_capacity.max()
+            if (max_deviation <= reop_tol[max_deviation.index]).all():
+                reop_converge = True
[…]
