### 2025-05-13 9af16b5
Set default line_loss_factor as non-zero
The line loss factor of zero leads to non-mutually exclusive forward and backward flows. Therefore, the line_loss_factor should be some small value.
