### 2025-05-14 2c10eb9
Fix daily hydropower modeling
