+++ src/pownet/reservoir/__init__.py
@@ -1,8 +1,6 @@
 """This is the reservoir module."""
 
 from .reservoir import Reservoir
-from .basin import Basin
-from .reservoir_operator import ReservoirOperator
 from .solve_release import (
     solve_release_from_target_storage,
     solve_release_from_dispatch,
