### 2025-05-14 d69abc8
Remove old reoperate method of Simulator class
