+++ src/pownet/core/simulation.py
@@ -260,63 +260,3 @@ class Simulator:
             lmp_df=self.system_record.get_lmp(),
             output_folder=output_folder,
         )
-
-    def reoperate(
-        self,
-        k: int,
-        builder: ModelBuilder,
-        init_conds: dict,
-        mip_gap: float = None,
-        timelimit: float = None,
-    ):
-        raise NotImplementedError("Reoperation is not implemented yet.")
-        reop_converge = False
-        reop_k = 0
-        while not reop_converge:
-            print(f"\nReservoirs reoperation iteration {reop_k}")
-            print("New Capacity vs. Current Dispatch")
-
-            # PowNet returns the hydropower dispatch in hourly resolution across the simulation horizon
-            hydro_dispatch, start_day, end_day = get_hydro_from_model(
-                self.model.model, k
-            )
-            # Convert to daily dispatch
-            hydro_dispatch = convert_to_daily_hydro(hydro_dispatch, start_day, end_day)
-            new_hydro_capacity = self.reservoir_operator.reoperate_basins(
-                pownet_dispatch=hydro_dispatch
-            )
-
-            for res in new_hydro_capacity.columns:
-                print(
-                    f"{res}: {round(new_hydro_capacity[res].sum(),2)} vs {round(hydro_dispatch[res].sum(),2)}",
-                )
-
-            max_deviation = (new_hydro_capacity - hydro_dispatch).abs().max()
-            # The tolerance for convergence should be 5% of the largest hydro capacity
-            reop_tol = 0.05 * new_hydro_capacity.max()
-            if (max_deviation <= reop_tol[max_deviation.index]).all():
-                reop_converge = True
-                print(f"PowNet: Day {k+1} - Reservoirs converged at iteration {reop_k}")
-
-            if reop_k > 50:
-                raise ValueError(
-                    "Reservoirs reoperation did not converge after 100 iterations"
-                )
-
-            # To reoptimize PowNet with the new hydropower capacity,
-            # update the builder class
-            builder.update_hydro_capacity(new_hydro_capacity)
-            self.model = builder.update(
-                k=k,
-                init_conds=init_conds,
-                mip_gap=mip_gap,
-                timelimit=timelimit,
-            )
-            self.model.optimize()
-
-            # Keep track of optimization time oand reoperation iterations
-            self.reop_opt_time += self.model.get_runtime()
-            reop_k += 1
-
-        # Record the number of iterations after convergence
-        self.reop_iter.append(reop_k)
