+++ docs/source/api_ref/pownet.optim_model.rst
@@ -1,0 +1,53 @@
+pownet.optim\_model package
+===========================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   pownet.optim_model.constraints
+
+Submodules
+----------
+
+pownet.optim\_model.model module
+--------------------------------
+
+.. automodule:: pownet.optim_model.model
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pownet.optim\_model.objfunc module
+----------------------------------
+
+.. automodule:: pownet.optim_model.objfunc
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pownet.optim\_model.rounding\_algo module
+-----------------------------------------
+
+.. automodule:: pownet.optim_model.rounding_algo
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pownet.optim\_model.variable module
+-----------------------------------
+
+.. automodule:: pownet.optim_model.variable
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pownet.optim_model
+   :members:
+   :undoc-members:
+   :show-inheritance:
