+++ docs/source/api_ref/pownet.reservoir.rst
@@ -1,0 +1,45 @@
+pownet.reservoir package
+========================
+
+Submodules
+----------
+
+pownet.reservoir.manager module
+-------------------------------
+
+.. automodule:: pownet.reservoir.manager
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pownet.reservoir.reservoir module
+---------------------------------
+
+.. automodule:: pownet.reservoir.reservoir
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pownet.reservoir.reservoir\_functions module
+--------------------------------------------
+
+.. automodule:: pownet.reservoir.reservoir_functions
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pownet.reservoir.solve\_release module
+--------------------------------------
+
+.. automodule:: pownet.reservoir.solve_release
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pownet.reservoir
+   :members:
+   :undoc-members:
+   :show-inheritance:
