+++ docs/source/api_ref/pownet.modeling.rst
@@ -1,45 +1,0 @@
-pownet.modeling package
-=======================
-
-Submodules
-----------
-
-pownet.modeling.constraint module
----------------------------------
-
-.. automodule:: pownet.modeling.constraint
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pownet.modeling.model module
-----------------------------
-
-.. automodule:: pownet.modeling.model
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pownet.modeling.objfunc module
-------------------------------
-
-.. automodule:: pownet.modeling.objfunc
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pownet.modeling.variable module
--------------------------------
-
-.. automodule:: pownet.modeling.variable
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: pownet.modeling
-   :members:
-   :undoc-members:
-   :show-inheritance:
