+++ src/test_pownet/test_core/test_builder.py
@@ -1,10 +1,15 @@
-""" Unit tests for the ModelBuilder class.
-"""
+"""Tests for the ModelBuilder class."""
 
 import os
 import unittest
-from pownet.core import ModelBuilder, SystemInput
-from pownet.folder_utils import get_model_dir
+from pownet import ModelBuilder, SystemInput
+from pownet.folder_utils import get_pownet_dir
+
+from pownet.builder.thermal import ThermalUnitBuilder
+from pownet.builder.hydro import HydroUnitBuilder
+from pownet.builder.nondispatch import NonDispatchUnitBuilder
+from pownet.builder.energy_storage import EnergyStorageUnitBuilder
+from pownet.builder.system import SystemBuilder
 
 
 class TestModelBuilder(unittest.TestCase):
@@ -13,7 +18,7 @@ class TestModelBuilder(unittest.TestCase):
     def setUp(self) -> None:
         # Load the test data
         test_model_library_path = os.path.abspath(
-            os.path.join(os.path.dirname(__file__), "..", "test_model_library")
+            os.path.join(get_pownet_dir(), "model_library")
         )
 
         self.inputs = SystemInput(
@@ -28,9 +33,18 @@ class TestModelBuilder(unittest.TestCase):
         self.model_builder = ModelBuilder(self.inputs)
 
     def test_init(self):
-        self.assertEqual(
-            self.model_builder.timesteps, range(1, self.inputs.sim_horizon + 1)
+        self.assertIsInstance(self.model_builder.thermal_builder, ThermalUnitBuilder)
+        self.assertIsInstance(self.model_builder.hydro_builder, HydroUnitBuilder)
+        self.assertIsInstance(
+            self.model_builder.nondispatch_builder, NonDispatchUnitBuilder
         )
+        self.assertIsInstance(
+            self.model_builder.storage_builder, EnergyStorageUnitBuilder
+        )
+        self.assertIsInstance(self.model_builder.system_builder, SystemBuilder)
+
+    def test_build(self):
+        pass
 
 
 if __name__ == "__main__":
