### 2025-05-15 cbb1d8f
Set the default hydropeak factor to 1
