### 2025-05-15 76cc2de
Use spill constraint in solve_release_from_dispatch
