+++ src/test_pownet/test_core/test_builder.py
@@ -3,7 +3,6 @@
 import os
 import unittest
 from pownet import ModelBuilder, SystemInput
-from pownet.folder_utils import get_pownet_dir
 
 from pownet.builder.thermal import ThermalUnitBuilder
 from pownet.builder.hydro import HydroUnitBuilder
@@ -11,15 +10,15 @@ from pownet.builder.nondispatch import NonDispatchUnitBuilder
 from pownet.builder.energy_storage import EnergyStorageUnitBuilder
 from pownet.builder.system import SystemBuilder
 
+from pownet.folder_utils import get_test_model_dir
+
 
 class TestModelBuilder(unittest.TestCase):
     """Unless otherwise stated, use "dummy_trade" over 24-hr as the test case."""
 
     def setUp(self) -> None:
         # Load the test data
-        test_model_library_path = os.path.abspath(
-            os.path.join(get_pownet_dir(), "model_library")
-        )
+        test_model_library_path = get_test_model_dir()
 
         self.inputs = SystemInput(
             input_folder=test_model_library_path,
