+++ src/test_pownet/test_model.py
@@ -1,10 +1,13 @@
-"""test_model.py"""
+""" This is test_model.py
+"""
 
+import os
 import unittest
+from unittest.mock import patch, MagicMock, call
 
 import gurobipy as gp
 import pandas as pd
-from pownet.optim_model import PowerSystemModel
+from pownet.modeling import PowerSystemModel
 
 
 class TestPowerSystemModel(unittest.TestCase):
