### 2025-05-15 3da33ef
Set default hydropeak factor to 1.0
