+++ pyproject.toml
@@ -8,6 +8,9 @@ where = ["src"]
 [tool.setuptools.package-data]
 pownet = ["database/*.csv"]
 
+[tool.coverage.run]
+disable_warnings = ["no-data-collected"]
+
 [project]
 name = "pownet"
 version = "2.2.0"
@@ -28,6 +31,7 @@ dependencies = [
     "statsmodels >= 0.14.4",
     "geopandas >= 1.0.1",
     "contextily >= 1.6.2",
+    "coverage >= 7.8.0",
 ]
 requires-python = ">=3.10"
 
