+++ pyproject.toml
@@ -31,7 +31,7 @@ dependencies = [
     "statsmodels >= 0.14.4",
     "geopandas >= 1.0.1",
     "contextily >= 1.6.2",
-    "coverage >= 7.8.0",
+    "coverage[toml] >= 7.8.0",
 ]
 requires-python = ">=3.10"
 
