+++ pyproject.toml
@@ -33,6 +33,7 @@ dependencies = [
     "geopandas >= 1.0.1",
     "contextily >= 1.6.2",
     "coverage[toml] >= 7.8.0",
+    "flake8 >= 7.2.0",
 ]
 requires-python = ">=3.10"
 
