+++ .github/workflows/python-app.yml
@@ -34,9 +34,9 @@ jobs:
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
-        flake8 . --ignore=E501,F401,W503,E266 --count --select=E9,F63,F7,F82 --show-source --statistics
+        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        flake8 . --ignore=E501,F401,W503,E266,C901 --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+        flake8 . --ignore=E501,W503,E266,C901 --count --exit-zero --max-complexity=10 --statistics
 
     - name: Unit testing
       run: |
