### 2025-05-15 78c027b
Add test_thermal_unit_constr.py
