+++ src/test_pownet/test_optim_model/test_constraints/test_thermal_unit_constr.py
@@ -1,0 +1,407 @@
+"""test_thermal_unit_constr.py"""
+
+import unittest
+import gurobipy as gp
+
+# Import the module to be tested
+from pownet.optim_model.constraints import thermal_unit_constr
+
+
+class TestThermalUnitConstraints(unittest.TestCase):
+
+    def setUp(self):
+        """Set up common resources for each test method."""
+        self.model = gp.Model("test_thermal_constraints")
+        self.model.setParam("OutputFlag", 0)
+
+        # Common test data
+        self.thermal_units = ["G1", "G2"]
+        self.sim_horizon_3 = 3
+        self.timesteps = range(1, self.sim_horizon_3 + 1)
+
+        # Define Gurobi variables
+        self.u_vars = self.model.addVars(
+            self.thermal_units, self.timesteps, vtype=gp.GRB.BINARY, name="u"
+        )
+        self.v_vars = self.model.addVars(
+            self.thermal_units, self.timesteps, vtype=gp.GRB.BINARY, name="v"
+        )
+        self.w_vars = self.model.addVars(
+            self.thermal_units, self.timesteps, vtype=gp.GRB.BINARY, name="w"
+        )
+        self.p_vars = self.model.addVars(
+            self.thermal_units, self.timesteps, vtype=gp.GRB.CONTINUOUS, name="p"
+        )
+        self.pbar_vars = self.model.addVars(
+            self.thermal_units, self.timesteps, vtype=gp.GRB.CONTINUOUS, name="pbar"
+        )
+        self.pthermal_vars = self.model.addVars(
+            self.thermal_units, self.timesteps, vtype=gp.GRB.CONTINUOUS, name="pthermal"
+        )
+        self.spin_vars = self.model.addVars(
+            self.thermal_units,
+            self.timesteps,
+            vtype=gp.GRB.CONTINUOUS,
+            lb=0,
+            name="spin",
+        )
+
+        self.initial_u_data = {"G1": 0, "G2": 1}
+        self.thermal_min_capacity_data = {"G1": 50, "G2": 100}
+
+        self.model.update()
+
+    def test_add_c_link_uvw_init(self):
+        # Add constraints for initial u values
+        constrs = thermal_unit_constr.add_c_link_uvw_init(
+            model=self.model,
+            u=self.u_vars,
+            v=self.v_vars,
+            w=self.w_vars,
+            thermal_units=self.thermal_units,
+            initial_u=self.initial_u_data,
+        )
+        self.model.update()
+
+        # Check the number of constraints added
+        self.assertEqual(len(constrs), len(self.thermal_units))
+
+        # Optional: Inspect a specific constraint
+        # Constraint name is "link_uvw_init[G1]" or "link_uvw_init[G2]"
+        g1_constr = self.model.getConstrByName("link_uvw_init[G1]")
+        self.assertIsNotNone(g1_constr)
+        self.assertEqual(g1_constr.Sense, gp.GRB.EQUAL)
+
+        # Expression: u[G1,1] - initial_u[G1] == v[G1,1] - w[G1,1]
+        # u[G1,1] - v[G1,1] + w[G1,1] == initial_u[G1]
+        row = self.model.getRow(g1_constr)
+        self.assertEqual(g1_constr.RHS, self.initial_u_data["G1"])
+
+        # Check coefficients
+        expected_coeffs_G1 = {
+            self.u_vars["G1", 1].VarName: 1.0,
+            self.v_vars["G1", 1].VarName: -1.0,
+            self.w_vars["G1", 1].VarName: 1.0,
+        }
+        actual_coeffs_G1 = {}
+        for i in range(row.size()):
[…]
