### 2025-05-15 e733073
Fix minor issues with unittest
