### 2025-05-16 17d1b76
Fix merging HydroCHIPPs
