+++ src/pownet/core/output.py
@@ -71,6 +71,21 @@ class OutputProcessor:
             node_variables["vartype"] == unit_type_map[unit_type]
         ].pivot(columns="node", index="hour", values="value")
 
+    def get_unit_hourly_generation(self, node_variables: pd.DataFrame) -> pd.DataFrame:
+        power_variables = self._get_power_variables(node_variables)
+        hourly_generation = (
+            power_variables[["unit", "value", "hour"]].groupby(["node", "hour"]).sum()
+        )
+        hourly_generation = hourly_generation.reset_index()
+        hourly_generation = hourly_generation.pivot(
+            columns=["hour"], index=["node"]
+        ).T.reset_index(drop=True)
+        # PowNet indexing starts at 1
+        hourly_generation.index += 1
+        hourly_generation.index.name = "Hour"
+
+        return hourly_generation
+
     def get_hourly_generation(self, node_variables: pd.DataFrame) -> pd.DataFrame:
         power_variables = self._get_power_variables(node_variables)
         hourly_generation = (
