### 2025-05-16 081ba2d
Update publications.rst
