### 2025-05-16 1600398
Update test_thermal_unit_constr.py
