+++ src/test_pownet/test_optim_model/test_constraints/test_thermal_unit_constr.py
@@ -2,6 +2,7 @@
 
 import unittest
 import gurobipy as gp
+import pandas as pd
 
 from pownet.optim_model.constraints import thermal_unit_constr
 
@@ -45,8 +46,20 @@ class TestThermalUnitConstrs(unittest.TestCase):
             name="spin",
         )
 
+        # --- Test data for constraints ---
+        # Initial conditions
         self.initial_u_data = {"G1": 0, "G2": 1}
+        self.initial_min_off_data = {
+            "G1": 5,  # initial_min_off > sim_horizon (5 > 3) => min_DT = 3
+            "G2": 2,  # initial_min_off <= sim_horizon (2 <= 3) => min_DT = 2
+        }
+
         self.thermal_min_capacity_data = {"G1": 50, "G2": 100}
+
+        self.thermal_derated_capacity = pd.DataFrame(
+            {"G1": [100, 100, 100], "G2": [200, 180, 160]},
+        )
+        self.thermal_derated_capacity.index += 1
 
         self.model.update()
 
@@ -343,9 +356,8 @@ class TestThermalUnitConstrs(unittest.TestCase):
         # Check constant on the RHS
         self.assertEqual(specific_constr.RHS, 0.0)
 
+        # Check the number of terms in the LHS
         row = self.model.getRow(specific_constr)
-
-        # Check number of non-zero coefficients in the LHS
         self.assertEqual(row.size(), 2)
 
         # Check which variables are involved in the LHS
@@ -355,6 +367,125 @@ class TestThermalUnitConstrs(unittest.TestCase):
             self.u_vars[test_unit, test_t].VarName,
         }
         self.assertSetEqual(actual_var_names, expected_var_names)
+        self.model.remove(constrs)
+
+    def test_add_c_link_pu_upper(self):
+        """Test the add_c_link_pu_upper constraint function."""
+        step_k_test = 1
+
+        constrs = thermal_unit_constr.add_c_link_pu_upper(
+            model=self.model,
+            pbar=self.pbar_vars,
+            u=self.u_vars,
+            timesteps=self.timesteps,
+            step_k=step_k_test,
+            thermal_units=self.thermal_units,
+            thermal_min_capacity=self.thermal_min_capacity_data,
+            thermal_derated_capacity=self.thermal_derated_capacity,
+        )
+        self.model.update()
+
+        # Check the number of constraints added
+        expected_num_constrs = len(self.thermal_units) * len(self.timesteps)
+        self.assertEqual(len(constrs), expected_num_constrs)
+
+        # Inspect a specific constraint (e.g., for G1 at timestep 1)
+        test_unit = self.thermal_units[0]  # "G1"
+        test_t = self.timesteps[0]  # 1
+
+        # Constraint name is "pthermal_ub[G1,1]"
+        constr_name = f"pthermal_ub[{test_unit},{test_t}]"
+        specific_constr = self.model.getConstrByName(constr_name)
+        self.assertIsNotNone(specific_constr, f"Constraint {constr_name} not found.")
+
+        # Original constraint:
+        # pbar[u,t] + min_cap[u]*u[u,t] <= derated_cap.loc[t+(k-1)*24, u]*u[u,t]
+        # Rearranged:
+        # pbar[u,t] + (min_cap[u] - derated_cap.loc[idx,u]) * u[u,t] <= 0
+        self.assertEqual(specific_constr.RHS, 0.0)
+
+        row = self.model.getRow(specific_constr)
[…]
