### 2025-05-16 fb44c5e
Rename and add tests to variable_func.py
Renamed "variable.py" to "variable_func.py"
