+++ src/test_pownet/test_optim_model/test_variable_func.py
@@ -1,0 +1,259 @@
+import unittest
+from unittest.mock import MagicMock, patch, call
+import pandas as pd
+
+from pownet.optim_model import variable_func
+
+
+class TestVariableFunctions(unittest.TestCase):
+
+    def setUp(self):
+        """Set up common test data and mocks."""
+        self.timesteps = range(
+            3
+        )  # Global timestep as per user request (implicitly via usage)
+        self.units = ["gen_A", "gen_B"]
+        self.edges = [("node1", "node2"), ("node2", "node3")]
+        self.step_k = 1  # Default step_k, can be overridden in specific tests
+
+        # Mock Gurobi model
+        self.mock_model = MagicMock()
+        # This mock will be returned by model.addVars
+        self.mock_vars_tupledict = MagicMock()
+        self.mock_model.addVars.return_value = self.mock_vars_tupledict
+
+        # Dummy capacity DataFrame for functions where its internal structure
+        # isn't directly tested by the function itself but passed to a mocked helper.
+        self.dummy_capacity_df = pd.DataFrame({"dummy_col": [1, 2, 3]})
+
+        # More structured capacity DataFrame for update_flow_vars
+        # Index needs to cover t + (step_k - 1) * 24
+        # For step_k=1, t=0,1,2 -> index 0,1,2
+        # For step_k=2, t=0,1,2 -> index 24,25,26
+        idx = pd.RangeIndex(start=0, stop=50, step=1)  # Sufficient for a few steps
+        data_for_flow_df = {
+            self.edges[0]: [100 + i for i in range(50)],
+            self.edges[1]: [200 + i for i in range(50)],
+        }
+        self.flow_capacity_df = pd.DataFrame(data_for_flow_df, index=idx)
+
+    @patch("pownet.optim_model.variable_func.GRB")
+    @patch("pownet.optim_model.variable_func.get_capacity_value")
+    def test_add_var_with_variable_ub(self, mock_get_capacity_value, mock_grb_module):
+        """Test the add_var_with_variable_ub function."""
+        var_name = "test_var"
+        step_k_test = 2
+
+        mock_grb_module.CONTINUOUS = "MOCK_GRB_CONTINUOUS_TYPE"
+
+        # Define a side effect for mock_get_capacity_value to simulate different capacities
+        def capacity_side_effect(t, unit, sk, df):
+            # sk is step_k, df is capacity_df
+            # Simple unique value based on inputs for verification
+            if unit == self.units[0]:
+                return 100 + t + sk
+            elif unit == self.units[1]:
+                return 200 + t + sk
+            return 0
+
+        mock_get_capacity_value.side_effect = capacity_side_effect
+
+        # Expected upper bounds dictionary
+        expected_ub_dict = {}
+        for t_val in self.timesteps:
+            for unit_val in self.units:
+                expected_ub_dict[(unit_val, t_val)] = capacity_side_effect(
+                    t_val, unit_val, step_k_test, self.dummy_capacity_df
+                )
+
+        # Call the function
+        created_vars = variable_func.add_var_with_variable_ub(
+            model=self.mock_model,
+            varname=var_name,
+            timesteps=self.timesteps,
+            step_k=step_k_test,
+            units=self.units,
+            capacity_df=self.dummy_capacity_df,
+        )
+
+        # Assert model.addVars was called correctly
+        self.mock_model.addVars.assert_called_once_with(
+            self.units,
+            self.timesteps,
+            lb=0,
+            ub=expected_ub_dict,
[…]
