### 2025-05-16 8dafb4a
Fix and add tests for rounding_algo.py
