### 2025-05-16 342b808
Fix test_variable_func.py
