### 2025-05-17 d383052
Add test_timeseries_model.py
