+++ src/test_pownet/test_stochastic/test_timeseries_model.py
@@ -1,0 +1,196 @@
+"""test_timeseries_model.py"""
+
+import unittest
+import pandas as pd
+from pownet.stochastic import timeseries_model
+
+
+class ConcreteTimeSeriesModel(timeseries_model.TimeSeriesModel):
+    """A concrete implementation for testing the abstract base class."""
+
+    def __init__(self):
+        super().__init__()
+        self._predictions = pd.Series(dtype=float)
+        self._residuals = pd.Series(dtype=float)
+        self._monthly_models = {}
+
+    @property
+    def monthly_models(self) -> dict:
+        return self._monthly_models
+
+    @property
+    def predictions(self) -> pd.Series:
+        return self._predictions
+
+    @property
+    def pred_residuals(self) -> pd.Series:
+        return self._residuals
+
+    def _fit(
+        self,
+        target_column: str,
+        arima_order: tuple[int, int, int],
+        seasonal_order: tuple[int, int, int, int],
+        exog_vars: list[str],
+    ) -> None:
+        # Minimal implementation for testing purposes
+        # In a real scenario, this would fit some model
+        self._monthly_models[1] = "dummy_model_for_month_1"
+        pass
+
+    def _predict(self) -> pd.Series:
+        # Minimal implementation
+        if not self.data.empty:
+            return pd.Series(
+                [1.0] * len(self.data), index=self.data.index, name="predictions"
+            )
+        return pd.Series(dtype=float)
+
+    def _get_synthetic(
+        self, exog_data: pd.DataFrame = None, seed: int = None
+    ) -> pd.Series:
+        # For this dummy, we'll keep the logic based on self.data, as it doesn't
+        # actually use the exogenous variables for its dummy output.
+        if not self.data.empty:
+            return pd.Series(
+                [0.5] * len(self.data), index=self.data.index, name="synthetic"
+            )
+        return pd.Series(dtype=float)
+
+    def _find_best_model(
+        self,
+        target_column: str,
+        exog_vars: list[str],
+        month_to_use: int,
+        seed: int,
+        suppress_warnings: bool,
+    ) -> tuple[tuple[int, int, int], tuple[int, int, int, int]]:
+        # Minimal implementation
+        return ((1, 0, 0), (0, 0, 0, 0))
+
+
+class TestTimeSeriesModel(unittest.TestCase):
+
+    def setUp(self):
+        self.model = ConcreteTimeSeriesModel()
+        self.sample_data = pd.DataFrame(
+            {
+                "datetime": pd.to_datetime(
+                    [
+                        "2023-01-01 00:00:00",
+                        "2023-01-01 01:00:00",
+                        "2023-01-01 02:00:00",  # Changed this line
+                        "2023-01-01 03:00:00",  # Added an extra point for more data
+                    ]
+                ),
+                "value": [10, 12, 15, 11],  # Adjusted values
+                "exog1": [1, 2, 3, 4],  # Adjusted exog
+            }
+        )
+        self.target_column = "value"
+
+    def test_initialization(self):
+        self.assertFalse(self.model._is_fitted)
+        self.assertFalse(self.model._is_loaded)
+        self.assertTrue(self.model.data.empty)
+        self.assertEqual(self.model.months, [])
+        self.assertIsNone(self.model.exog_vars)
+
+    def test_load_data_success(self):
+        self.model.load_data(self.sample_data.copy())
+        self.assertTrue(self.model._is_loaded)
+        self.assertFalse(self.model.data.empty)
+        self.assertIn(pd.Timestamp("2023-01-01 00:00:00"), self.model.data.index)
+        self.assertEqual(self.model.data.index.freqstr, "h")
[…]
