### 2025-05-17 006481d
Update reservoir.py
