+++ src/pownet/folder_utils.py
@@ -1,3 +1,5 @@
+"folder_utils.py: Folder utility functions for pownet package." ""
+
 import os
 
 
