+++ src/test_pownet/test_coupler.py
@@ -1,0 +1,299 @@
+"test_coupler.py"
+
+import unittest
+from unittest.mock import MagicMock, PropertyMock, call
+import logging
+
+# Assuming coupler.py is in the same directory or accessible via PYTHONPATH
+from pownet.coupler import PowerWaterCoupler
+
+# For type hinting if needed
+from pownet import ModelBuilder as ActualModelBuilder
+from pownet.reservoir import ReservoirManager as ActualReservoirManager
+
+# Disable logging for cleaner test output
+logging.disable(logging.CRITICAL)
+
+
+class TestPowerWaterCoupler(unittest.TestCase):
+
+    def setUp(self):
+        """Common setup for all tests."""
+        self.mock_model_builder = MagicMock(spec=ActualModelBuilder)
+        self.mock_reservoir_manager = MagicMock(spec=ActualReservoirManager)
+
+        # --- Configure ModelBuilder Mock ---
+        mock_mb_inputs = MagicMock()
+        mock_mb_inputs.sim_horizon = 48  # Allows num_days_in_step = 2
+        type(self.mock_model_builder).inputs = PropertyMock(return_value=mock_mb_inputs)
+
+        self.mock_power_system_model = MagicMock()
+        self.mock_power_system_model.get_runtime.return_value = 0.5
+        self.mock_model_builder.update_daily_hydropower_capacity.return_value = (
+            self.mock_power_system_model
+        )
+        self.mock_model_builder.get_phydro.return_value = {}
+
+        # --- Configure ReservoirManager Mock ---
+        self.mock_reservoir_manager.simulation_order = ["H1", "H2"]
+        self.mock_reservoir_manager.reoperate.return_value = {}
+
+        self.coupler = PowerWaterCoupler(
+            model_builder=self.mock_model_builder,
+            reservoir_manager=self.mock_reservoir_manager,
+            solver="test_solver",
+            mip_gap=0.001,
+            timelimit=300,
+            log_to_console=False,
+        )
+        # num_days_in_step should be 2 based on sim_horizon = 48
+        self.assertEqual(self.coupler.num_days_in_step, 2)
+
+    def _create_mock_dispatch_var(self, value):
+        mock_var = MagicMock()
+        mock_var.X = value
+        return mock_var
+
+    def test_initialization(self):
+        self.assertIs(self.coupler.model_builder, self.mock_model_builder)
+        self.assertEqual(self.coupler.num_days_in_step, 2)
+        self.assertEqual(self.coupler.reop_iter, [])
+        self.assertEqual(self.coupler.reop_opt_time, 0.0)
+
+    def test_getters(self):
+        self.coupler.reop_opt_time = 10.5
+        self.assertEqual(self.coupler.get_reop_opt_time(), 10.5)
+        self.coupler.reop_iter = [1, 2, 3]
+        self.assertEqual(self.coupler.get_reop_iter(), [1, 2, 3])
+
+    def test_reoperate_converges_immediately_multi_day(self):
+        step_k = 10  # 1-indexed global start day
+        # num_days_in_step is 2, so days_in_step = range(10, 12) -> global days 10, 11
+
+        # Hydropower dispatch from ModelBuilder (hourly, spanning 48 hours)
+        # varname[1] is 1-indexed hour in ModelBuilder's 48h horizon
+        initial_phydro = {
+            # Day step_k (Global Day 10)
+            ("H1", 1): self._create_mock_dispatch_var(1.0),  # Hour 1, Day 10
+            ("H1", 24): self._create_mock_dispatch_var(1.0),  # Hour 24, Day 10
+            ("H2", 1): self._create_mock_dispatch_var(0.5),  # Hour 1, Day 10
[…]
