+++ src/test_pownet/test_optim_model/test_constraints/test_system_constr.py
@@ -1,0 +1,1409 @@
+"""test_system_constr.py: Unittests for system-wide constraints."""
+
+import math
+import unittest
+
+import gurobipy as gp
+import pandas as pd
+
+from pownet.optim_model.constraints import system_constr
+
+
+class TestSystemConstraints(unittest.TestCase):
+    """Test cases for system-wide constraints."""
+
+    def setUp(self):
+        """Set up common resources for each test method."""
+        self.model = gp.Model("test_system_constraints")
+        self.model.setParam("OutputFlag", 0)  # Suppress Gurobi output
+
+        self.nodes = ["N1", "N2", "N3"]
+        self.thermal_units = ["G1", "G2"]
+        self.storage_units = ["S1"]
+        self.edges = [
+            ("N1", "N2"),
+            ("N1", "N3"),
+            ("N2", "N3"),
+        ]  # List of network edges (branches)
+        self.max_demand_node_name = "N1"
+
+        self.sim_horizon = 3
+        self.timesteps = range(1, self.sim_horizon + 1)  # Timesteps: 1, 2, 3
+
+        # Corresponds to the '24' in susceptance lookup: t + (step_k - 1) * 24
+        self.hours_per_step = 24
+
+        # Variables for add_c_reserve_req_1
+        self.spin_vars = self.model.addVars(
+            self.thermal_units,
+            self.timesteps,
+            vtype=gp.GRB.CONTINUOUS,
+            lb=0,
+            name="spin",
+        )
+
+        # Variables for add_c_reserve_req_2
+        self.pbar_vars = self.model.addVars(
+            self.thermal_units,
+            self.timesteps,
+            vtype=gp.GRB.CONTINUOUS,
+            lb=0,
+            name="pbar",
+        )
+        self.u_vars = self.model.addVars(
+            self.thermal_units, self.timesteps, vtype=gp.GRB.BINARY, name="u"
+        )
+        self.thermal_min_capacity_data = {"G1": 50.0, "G2": 30.0}
+
+        # Common variables for reserve constraints
+        self.charge_state_vars = self.model.addVars(
+            self.storage_units,
+            self.timesteps,
+            vtype=gp.GRB.CONTINUOUS,
+            lb=0,
+            name="charge_state",
+        )
+        self.spin_shortfall_vars = self.model.addVars(
+            self.timesteps, vtype=gp.GRB.CONTINUOUS, lb=0, name="spin_shortfall"
+        )
+
+        # Variables for add_c_link_shortfall
+        self.load_shortfall_vars = self.model.addVars(
+            self.timesteps,
+            vtype=gp.GRB.CONTINUOUS,
+            name="load_shortfall",
+            lb=-gp.GRB.INFINITY,
+        )
+        self.p_pmismatch_vars = self.model.addVars(
+            self.timesteps, vtype=gp.GRB.CONTINUOUS, lb=0, name="p_pmismatch"
+        )
+        self.n_pmismatch_vars = self.model.addVars(
+            self.timesteps, vtype=gp.GRB.CONTINUOUS, lb=0, name="n_pmismatch"
+        )
+
+        # Variables for voltage angle constraints (ref_node and angle_diff)
+        self.theta_vars = self.model.addVars(
+            self.nodes,
+            self.timesteps,
+            vtype=gp.GRB.CONTINUOUS,
+            name="theta",
+            lb=-math.pi / 2,
+            ub=math.pi / 2,
+        )
+
+        # Variables for add_c_angle_diff
+        self.flow_fwd_vars = gp.tupledict()
+        self.flow_bwd_vars = gp.tupledict()
+        for a_node, b_node in self.edges:
+            for t_step in self.timesteps:
[…]

[…]
