### 2025-05-18 0eb24fb
Update variable_func.py
