+++ src/test_pownet/test_optim_model/test_variable_func.py
@@ -1,30 +1,31 @@
-"""test_variable_func.py"""
+"""test_variable_func.py: Unit tests for variable_func.py."""
 
 import unittest
 from unittest.mock import MagicMock, patch, call
 import pandas as pd
 
+# Assuming variable_func is in pownet.optim_model directory
+# Adjust the import path if your directory structure is different.
+# For example, if pownet is in your PYTHONPATH:
 from pownet.optim_model import variable_func
+
+# If variable_func.py is in the same directory as the test for local testing,
+# you might use:
+# import variable_func
 
 
 class TestVariableFunctions(unittest.TestCase):
 
-    @classmethod
-    def setUpClass(cls):
-        """Set up class-level resources or patches before any tests in the class run."""
-        variable_func.VAR_PREFIX_THERMAL_GENERATION = "thermal_generation"
+    # Removed setUpClass as VAR_PREFIX_THERMAL_GENERATION is no longer used
+    # in variable_func.py
 
     def setUp(self):
         """Set up common test data and mocks."""
 
-        self.VAR_PREFIX_THERMAL_GENERATION = "thermal_generation"
-
-        self.timesteps = range(
-            3
-        )  # Global timestep as per user request (implicitly via usage)
+        self.timesteps = range(3)
         self.units = ["gen_A", "gen_B"]
         self.edges = [("node1", "node2"), ("node2", "node3")]
-        self.step_k = 1  # Default step_k, can be overridden in specific tests
+        self.step_k = 1  # Default step_k, can be overridden
 
         # Mock Gurobi model
         self.mock_model = MagicMock()
@@ -37,9 +38,6 @@ class TestVariableFunctions(unittest.TestCase):
         self.dummy_capacity_df = pd.DataFrame({"dummy_col": [1, 2, 3]})
 
         # More structured capacity DataFrame for update_flow_vars
-        # Index needs to cover t + (step_k - 1) * 24
-        # For step_k=1, t=0,1,2 -> index 0,1,2
-        # For step_k=2, t=0,1,2 -> index 24,25,26
         idx = pd.RangeIndex(start=0, stop=50, step=1)  # Sufficient for a few steps
         data_for_flow_df = {
             self.edges[0]: [100 + i for i in range(50)],
@@ -54,12 +52,12 @@ class TestVariableFunctions(unittest.TestCase):
         var_name = "test_var"
         step_k_test = 2
 
+        # Mock the GRB constant
         mock_grb_module.CONTINUOUS = "MOCK_GRB_CONTINUOUS_TYPE"
 
-        # Define a side effect for mock_get_capacity_value to simulate different capacities
+        # Define a side effect for mock_get_capacity_value
         def capacity_side_effect(t, unit, sk, df):
-            # sk is step_k, df is capacity_df
-            # Simple unique value based on inputs for verification
+            self.assertIs(df, self.dummy_capacity_df)  # Ensure correct df is passed
             if unit == self.units[0]:
                 return 100 + t + sk
             elif unit == self.units[1]:
@@ -68,7 +66,7 @@ class TestVariableFunctions(unittest.TestCase):
 
         mock_get_capacity_value.side_effect = capacity_side_effect
 
-        # Expected upper bounds dictionary
+        # Expected upper bounds dictionary based on the side effect
         expected_ub_dict = {}
         for t_val in self.timesteps:
             for unit_val in self.units:
@@ -92,7 +90,7 @@ class TestVariableFunctions(unittest.TestCase):
             self.timesteps,
             lb=0,
             ub=expected_ub_dict,
-            vtype="MOCK_GRB_CONTINUOUS_TYPE",  # Check if the mocked GRB type is used
[…]
