### 2025-05-18 6eb80c3
Revise hydropower formulation
Removed binary variables from the default formulation to speed up the computation.
