+++ src/pownet/builder/hydro.py
@@ -1,21 +1,28 @@
-"""hydro.py: Hydro unit builder."""
+"""hydro.py: Hydro unit builder. This includes hourly, daily, and weekly constraints for hydro units."""
 
 from .basebuilder import ComponentBuilder
 
 import gurobipy as gp
 
 from ..input import SystemInput
+from ..optim_model.variable_func import (
+    add_var_with_variable_ub,
+    update_var_with_variable_ub,
+)
 from ..optim_model.objfunc import get_marginal_cost_coeff
 from ..optim_model.constraints import nondispatch_constr
 
 
 class HydroUnitBuilder(ComponentBuilder):
-    """Builder class for hydro units.
+    """Builder class for hydro units. If a hydro unit's hourly availability is provided,
+    then the hydro unit's dispatch variables are limited by them. If a hydro unit's
+    hourly availability is not provided, as in the case with daily/weekly hydro units,
+    then the hydro unit's dispatch variables are limited by the contracted capacity.
 
     Variables
     ===========================
     - `phydro`: Hydropower output. Unit: MW.
-    - `uhydro`: Hydropower unit status. Unit: binary (0 or 1).
+    - `uhydro`: Hydropower unit status. Unit: binary (0 or 1). (optional)
 
     Fixed objective terms
     ===========================
@@ -27,8 +34,8 @@ class HydroUnitBuilder(ComponentBuilder):
 
     Constraints
     ===========================
+    - Hourly/Daily/Weekly hydropower limits
     - Linking upper bounds of dispatch variables to unit status variables
-    - Daily hydropower limits
 
     """
 
@@ -37,6 +44,10 @@ class HydroUnitBuilder(ComponentBuilder):
 
         # Variables
         self.phydro = gp.tupledict()
+        self.hourly_phydro = gp.tupledict()
+        self.daily_phydro = gp.tupledict()
+        self.weekly_phydro = gp.tupledict()
+
         self.uhydro = gp.tupledict()
 
         # Fixed objective terms
@@ -46,11 +57,11 @@ class HydroUnitBuilder(ComponentBuilder):
         self.total_energy_cost_expr = gp.LinExpr()
 
         # Constraints
-        self.c_link_hydro_pu = gp.tupledict()
-        self.c_link_weekly_hydro_pu = gp.tupledict()
-
+        self.c_hourly_hydro_ub = gp.tupledict()
         self.c_hydro_limit_daily = gp.tupledict()
         self.c_hydro_limit_weekly = gp.tupledict()
+
+        self.c_link_hydro_pu = gp.tupledict()
 
     def add_variables(self, step_k: int) -> None:
         """Add variables to the model for hydro units.
@@ -58,26 +69,56 @@ class HydroUnitBuilder(ComponentBuilder):
         Args:
             step_k (int): Current time step.
         """
-        self.phydro = self.model.addVars(
-            self.inputs.hydro_units,
+        # --- Hourly hydropower
+        self.hourly_phydro = add_var_with_variable_ub(
+            model=self.model,
+            varname="phydro",
+            timesteps=self.timesteps,
+            step_k=step_k,
+            units=self.inputs.hydro_unit_node.keys(),
+            capacity_df=self.inputs.hydro_capacity,
+        )
+
+        # --- Daily/weekly hydropower are limited by contracted capacity
+        self.daily_phydro = self.model.addVars(
+            self.inputs.daily_hydro_unit_node.keys(),
             self.timesteps,
             lb=0,
             ub={
                 (unit, t): self.inputs.hydro_contracted_capacity[unit]
+                for unit in self.inputs.daily_hydro_unit_node.keys()
                 for t in self.timesteps
-                for unit in self.inputs.hydro_units
             },
[…]
