### 2025-05-18 7302bbf
Create test_energy_storage_constr.py
