+++ src/test_pownet/test_optim_model/test_constraints/test_energy_storage_constr.py
@@ -1,0 +1,368 @@
+import unittest
+import gurobipy as gp
+
+
+from pownet.optim_model.constraints import energy_storage_constr
+
+
+class TestEnergyStorageConstraints(unittest.TestCase):
+    """Test cases for energy storage unit constraints."""
+
+    def setUp(self):
+        """Set up common resources for each test method."""
+        self.model = gp.Model("test_energy_storage_constraints")
+        self.model.setParam("OutputFlag", 0)  # Suppress Gurobi output
+
+        self.timesteps = range(1, 4)  # Timesteps: 1, 2, 3
+        self.sim_horizon = 3  # Corresponds to the end of self.timesteps
+        self.storage_units = ["ESS1", "ESS2"]
+
+        self.max_charge_cap = {
+            unit: 50.0 + idx * 10 for idx, unit in enumerate(self.storage_units)
+        }
+        self.max_discharge_cap = {
+            unit: 40.0 + idx * 10 for idx, unit in enumerate(self.storage_units)
+        }
+
+        self.charge_state_init = {
+            unit: 100.0 + idx * 20 for idx, unit in enumerate(self.storage_units)
+        }
+        self.charge_efficiency = {
+            unit: 0.95 - idx * 0.02 for idx, unit in enumerate(self.storage_units)
+        }
+        # Ensure discharge_efficiency is not zero
+        self.discharge_efficiency = {
+            unit: 0.90 - idx * 0.02 for idx, unit in enumerate(self.storage_units)
+        }
+        self.self_discharge_rate = {
+            unit: 0.01 + idx * 0.005 for idx, unit in enumerate(self.storage_units)
+        }
+
+        # Variables
+        self.pcharge = self.model.addVars(
+            self.storage_units, self.timesteps, vtype=gp.GRB.CONTINUOUS, name="pcharge"
+        )
+        self.ucharge = self.model.addVars(
+            self.storage_units, self.timesteps, vtype=gp.GRB.BINARY, name="ucharge"
+        )
+        self.pdischarge = self.model.addVars(
+            self.storage_units,
+            self.timesteps,
+            vtype=gp.GRB.CONTINUOUS,
+            name="pdischarge",
+        )
+        self.udischarge = self.model.addVars(
+            self.storage_units, self.timesteps, vtype=gp.GRB.BINARY, name="udischarge"
+        )
+        self.charge_state = self.model.addVars(
+            self.storage_units,
+            self.timesteps,
+            vtype=gp.GRB.CONTINUOUS,
+            name="charge_state",
+        )
+
+        self.model.update()
+
+    def tearDown(self):
+        """Clean up resources after each test method."""
+        self.model.dispose()
+
+    def _check_constr_details(
+        self,
+        constr_name_base,
+        constr_dict,
+        expected_keys_iterable,
+        expected_sense,
+        get_expected_rhs_func,
+        get_expected_lhs_coeffs_func,
+    ):
+        """Helper to check constraint details."""
+        self.assertEqual(len(constr_dict), len(list(expected_keys_iterable)))
+
+        for key_tuple in expected_keys_iterable:
+            # Construct the full constraint name as Gurobi creates it
+            if isinstance(key_tuple, str):  # Single key for some constraints
+                full_constr_name = f"{constr_name_base}[{key_tuple}]"
+                key_for_data = key_tuple  # used for fetching data from dicts
+            else:  # Multiple keys (unit, t)
+                full_constr_name = (
+                    f"{constr_name_base}[{','.join(map(str, key_tuple))}]"
+                )
+                key_for_data = key_tuple
+
+            current_constr = self.model.getConstrByName(full_constr_name)
+            self.assertIsNotNone(
+                current_constr, f"Constraint {full_constr_name} not found."
+            )
+
[…]
