+++ src/pownet/optim_model/constraints/energy_storage_constr.py
@@ -1,7 +1,6 @@
 """energy_storage_constr.py: Constraints for energy storage units."""
 
 import gurobipy as gp
-import pandas as pd
 
 
 def add_c_link_ess_charge(
