+++ src/pownet/reservoir/reservoir_functions.py
@@ -20,7 +20,7 @@ def find_upstream_units(flow_paths: pd.DataFrame, unit_name: str) -> list[str]:
 
 def find_downstream_flow_fractions(
     flow_paths: pd.DataFrame, unit_name: str
-) -> list[str, float]:
+) -> dict[str, float]:
     """Find downstream units for a given unit.
 
     Args:
@@ -28,7 +28,7 @@ def find_downstream_flow_fractions(
         unit_name (str): The name of the unit to find downstream units for.
 
     Returns:
-        list[str, float]: List of downstream unit names and their flow fractions.
+        dict[str, float]: Dict of downstream unit names and their flow fractions.
     """
     return (
         flow_paths.loc[flow_paths["source"] == unit_name, ["sink", "flow_fraction"]]
@@ -54,8 +54,7 @@ def find_simulation_order(
     G = nx.DiGraph(edgelist)
     try:
         simulation_order = list(nx.topological_sort(G))
-        # The current list excludes reservoirs that are not in the flow paths,
-        # so we need to add them to the end of the list
+        # Some reservoirs are not in the flow paths, so we need to add them to the end of the list
         reservoirs_not_in_paths = []
         for reservoir in reservoir_names:
             if reservoir not in simulation_order:
@@ -78,8 +77,8 @@ def adjust_hydropeaking(
     Also, the release cannot be lower than the minimum release or higher than the maximum release.
 
     Args:
-        release (float): The release for the current day. Can be both negative and positive
-        release_t0 (float): The release for the previous day. Can be both negative and positive
+        release (float): The release for the current day.
+        release_t0 (float): The release for the previous day.
         max_release (float): The maximum release as a positive value
         min_release (float): The minimum release as a positive value
         hydropeak_factor (float): The factor to limit the change in release. Default is 0.15.
@@ -112,78 +111,78 @@ def adjust_hydropeaking(
     return adj_release
 
 
-def calc_min_environ_flow(
-    inflow: float,
-    mean_annual_flow: float,
-    max_release: float,
-) -> float:
-    """Tdetermine the minimum amount of water that should be released
-    from a reservoir to maintain the health of the downstream ecosystem.
+# def calc_min_environ_flow(
+#     inflow: float,
+#     mean_annual_flow: float,
+#     max_release: float,
+# ) -> float:
+#     """Tdetermine the minimum amount of water that should be released
+#     from a reservoir to maintain the health of the downstream ecosystem.
 
-    Example of three cases:
+#     Example of three cases:
 
-    1) If the inflow is less than 40% of the mean annual flow,
-    the minimum flow is set to 60% of the inflow.
-    This ensures that a reasonable portion of the limited water
-    is still released to support the ecosystem.
+#     1) If the inflow is less than 40% of the mean annual flow,
+#     the minimum flow is set to 60% of the inflow.
+#     This ensures that a reasonable portion of the limited water
+#     is still released to support the ecosystem.
 
-    2) If the inflow is greater than 80% of the mean annual flow,
-    the minimum flow is 30% of the inflow. a smaller percentage is
-    released since the ecosystem is likely receiving ample water already.
+#     2) If the inflow is greater than 80% of the mean annual flow,
+#     the minimum flow is 30% of the inflow. a smaller percentage is
+#     released since the ecosystem is likely receiving ample water already.
 
-    3) Otherwise, the minimum flow is 45% of the inflow.
+#     3) Otherwise, the minimum flow is 45% of the inflow.
 
-    Args:
[…]
