### 2025-05-18 1da3ec6
Refactor and add test for ModelBuilder
