+++ src/pownet/core/__init__.py
@@ -1,6 +1,6 @@
 """This is the core module."""
 
-from .builder import ModelBuilder
+from .model_builder import ModelBuilder
 from .output import OutputProcessor
 from .visualizer import Visualizer
 from .record import SystemRecord
