+++ src/test_pownet/test_core/test_model_builder.py
@@ -1,0 +1,436 @@
+"""test_builder.py: Unit tests for the ModelBuilder class."""
+
+import unittest
+from unittest.mock import MagicMock, patch, call
+
+from pownet import ModelBuilder
+
+from gurobipy import GRB  # Used directly for GRB.MINIMIZE
+
+
+# Custom Mock for gurobipy.LinExpr to make objective summation testable
+class MockLinExpr:
+    def __init__(self, value=0.0, name=""):
+        self.value = float(value)
+        self._name = name  # For debugging
+
+    def __add__(self, other):
+        other_value = other.value if isinstance(other, MockLinExpr) else float(other)
+        return MockLinExpr(self.value + other_value, f"{self._name}_add")
+
+    def __radd__(self, other):  # For sum() or when LinExpr is on the right
+        other_value = other.value if isinstance(other, MockLinExpr) else float(other)
+        return MockLinExpr(self.value + other_value, f"{self._name}_radd")
+
+    def __iadd__(self, other):
+        other_value = other.value if isinstance(other, MockLinExpr) else float(other)
+        self.value += other_value
+        return self
+
+    def copy(self):
+        return MockLinExpr(self.value, f"{self._name}_copy")
+
+    def getValue(self):
+        return self.value
+
+    def __repr__(self):
+        return f"MockLinExpr({self.value}, name='{self._name}')"
+
+
+PATCH_BASE = "pownet.core.model_builder"
+
+
+# The order of decorators matters: they are applied from bottom to top.
+# The mock for the bottom-most decorator is the first argument to the test method.
+@patch(f"{PATCH_BASE}.SystemBuilder", autospec=True)
+@patch(f"{PATCH_BASE}.EnergyStorageUnitBuilder", autospec=True)
+@patch(f"{PATCH_BASE}.NonDispatchUnitBuilder", autospec=True)
+@patch(f"{PATCH_BASE}.HydroUnitBuilder", autospec=True)
+@patch(f"{PATCH_BASE}.ThermalUnitBuilder", autospec=True)
+@patch(
+    f"{PATCH_BASE}.PowerSystemModel", autospec=True
+)  # This is from pownet.optim_model but imported into PATCH_BASE
+@patch(
+    f"{PATCH_BASE}.gp.LinExpr",
+    side_effect=lambda name_suffix="": MockLinExpr(
+        0.0, f"initial_LinExpr_{name_suffix}"
+    ),
+)
+@patch(f"{PATCH_BASE}.gp.Model", autospec=True)  # gp is an alias in PATCH_BASE
+@patch(
+    f"{PATCH_BASE}.SystemInput", autospec=True
+)  # SystemInput is from ..input but imported into PATCH_BASE
+class TestModelBuilder(unittest.TestCase):
+
+    # setUp is not strictly needed here as mocks are passed directly via decorators.
+    # def setUp(self):
+    #     pass
+
+    def _configure_mock_builder(
+        self, mock_builder_instance, name="test", fixed_obj_val=1.0, var_obj_val=10.0
+    ):
+        """Helper to configure a mock sub-builder."""
+        mock_builder_instance.get_fixed_objective_terms.return_value = MockLinExpr(
+            fixed_obj_val, f"{name}_fixed_obj"
+        )
+        mock_builder_instance.get_variable_objective_terms.return_value = MockLinExpr(
+            var_obj_val, f"{name}_var_obj"
+        )
+
+        # Mock attributes that SystemBuilder.add_constraints might need
+        mock_builder_instance.spin = MagicMock(name=f"{name}_spin_var")
+        mock_builder_instance.vpowerbar = MagicMock(name=f"{name}_vpowerbar_var")
[…]
