+++ src/test_pownet/test_builder/test_basebuilder.py
@@ -1,0 +1,190 @@
+"""test_basebuilder.py: Unit tests for the ComponentBuilder abstract base class."""
+
+import unittest
+from unittest.mock import MagicMock, patch
+
+# Import ComponentBuilder from its actual location
+from pownet.builder.basebuilder import ComponentBuilder
+
+# PATCH_BASE should be the module where ComponentBuilder is defined,
+# as this is the context where its internal imports (like gp and SystemInput) are resolved.
+PATCH_BASE = "pownet.builder.basebuilder"
+
+# We need ABC for creating test subclasses
+from abc import ABC, abstractmethod  # ABC is implicitly used by ComponentBuilder
+
+
+# Minimal concrete implementation for testing ComponentBuilder
+class MinimalConcreteBuilder(ComponentBuilder):
+    """A minimal concrete subclass for testing ComponentBuilder."""
+
+    def add_variables(self, step_k: int) -> None:
+        """Mock implementation."""
+        pass
+
+    def get_fixed_objective_terms(self) -> MagicMock:  # Actual type is gp.LinExpr
+        """Mock implementation."""
+        return MagicMock(name="MockLinExpr_fixed")
+
+    def get_variable_objective_terms(
+        self, step_k: int, **kwargs
+    ) -> MagicMock:  # Actual type is gp.LinExpr
+        """Mock implementation."""
+        return MagicMock(name="MockLinExpr_variable")
+
+    def add_constraints(self, step_k: int, init_conds: dict, **kwargs) -> None:
+        """Mock implementation."""
+        pass
+
+    def update_variables(self, step_k: int) -> None:
+        """Mock implementation."""
+        pass
+
+    def update_constraints(self, step_k: int, init_conds: dict, **kwargs) -> None:
+        """Mock implementation."""
+        pass
+
+    def get_variables(
+        self,
+    ) -> dict[str, MagicMock]:  # Actual type is dict[str, gp.tupledict]
+        """Mock implementation."""
+        return {"mock_var": MagicMock(name="MockTupleDict")}
+
+
+# An incomplete subclass for testing abstract method enforcement
+class IncompleteBuilder(ComponentBuilder):
+    """An incomplete subclass that misses some abstract methods."""
+
+    def add_variables(self, step_k: int) -> None:
+        pass
+
+    def get_fixed_objective_terms(self) -> MagicMock:
+        return MagicMock(name="MockLinExpr_fixed_incomplete")
+
+    # Missing: get_variable_objective_terms, add_constraints, etc.
+    def get_variables(self) -> dict[str, MagicMock]:
+        return {"mock_var_incomplete": MagicMock(name="MockTupleDict_incomplete")}
+
+    # To make it instantiable for other tests, we would need to implement all other abstract methods.
+    # For this test, we want it to remain abstract.
+    @abstractmethod  # Explicitly mark remaining methods as abstract if not implemented
+    def get_variable_objective_terms(self, step_k: int, **kwargs) -> MagicMock:
+        pass
+
+    @abstractmethod
+    def add_constraints(self, step_k: int, init_conds: dict, **kwargs) -> None:
+        pass
+
+    @abstractmethod
+    def update_variables(self, step_k: int) -> None:
+        pass
+
+    @abstractmethod
+    def update_constraints(self, step_k: int, init_conds: dict, **kwargs) -> None:
+        pass
+
+
+@patch(f"{PATCH_BASE}.SystemInput", autospec=True)
+@patch(f"{PATCH_BASE}.gp")  # Patch the 'gp' alias used in basebuilder.py
+class TestComponentBuilder(unittest.TestCase):
+
+    def _configure_mock_gp_alias(self, mock_gp_alias):
[…]
