### 2025-05-19 20e3874
docs: add code of conduct
