### 2025-05-19 596d937
docs: Create CONTRIBUTING.md
