+++ docs/source/api_ref/pownet.data_model.rst
@@ -1,21 +1,9 @@
 pownet.data\_model package
 ==========================
 
-Submodules
-----------
-
 pownet.data\_model.reservoir module
 -----------------------------------
 
 .. automodule:: pownet.data_model.reservoir
    :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: pownet.data_model
-   :members:
-   :undoc-members:
-   :show-inheritance:
+   :no-index:
