+++ docs/source/api_ref/modules.rst
@@ -1,7 +1,0 @@
-pownet
-======
-
-.. toctree::
-   :maxdepth: 4
-
-   pownet
