+++ docs/source/reference/contributing.rst
@@ -1,0 +1,10 @@
+
+.. autosummary::
+    :toctree: _source/
+
+################################
+Contributing
+################################
+
+    
+.. mdinclude:: ../../.github/CONTRIBUTING.md
