+++ docs/requirements.txt
@@ -1,5 +1,6 @@
 sphinx >= 8.1.0
 sphinx-rtd-theme >= 3.0.2
+sphinx-mdinclude >= 0.6.2
 nbsphinx >= 0.9.7
 nbsphinx_link >= 1.3.1
 sphinxcontrib-bibtex >= 2.6.3
