### 2025-05-19 Critical-Infrastructure-Systems-Lab/PowNet#21 @phumthep
I have added a code of conduct page and a contributing guideline page with 596d937.

Contributing guideline can also be found at
https://pownet.readthedocs.io/en/latest/reference/contributing.html
