### 2025-05-19 5729b2e
docs: Revise joss paper
