+++ joss_submission/paper.md
@@ -13,48 +13,58 @@ authors:
   - name: Hisham Eldardiry
     orcid: 0000-0002-2932-7459
     affiliation: 1
+  - name: Matija	Pavičević
+    Orcid: 0000-0002-6524-5581
+    affiliation: 2
+  - name: Jerry Zhuoer Feng
+    orcid: 0009-0004-5129-1167
+    affiliation: 1
   - name: Stefano Galelli
     orcid: 0000-0003-2316-3243
     affiliation: 1
 affiliations:
  - name: Cornell University, USA
    index: 1
+ - name: Argonne National Laboratory, USA
+   index: 2
 date: 6 October 2024
 bibliography: paper.bib
 ---
 
+
 # Summary
 
-Managing the complex network of power stations and transmission lines that deliver electricity across large spatial domains (e.g., country and continental scale) requires a variety of mathematical models, most of which are formulated as mathematical optimization problems. Production Cost Models (PCM) are among the most common types of models found in the power systems literature; their task is for planning the short-term operation of generators [@garver1962threebin]. In contrast to other power system models, PCMs thus employ a simpler representation of power flow and do not account for long-term investment decisions for system components. `PowNet 2.0` is a Python-based production cost modeling framework that simulates and analyzes the most cost-effective way of meeting electricity demand using various electricity sources (coal, natural gas, and renewables) such that sufficient electricity is generated at each hour while considering factors like fuel prices, availability of renewables, and regulatory requirements.
+Managing the complex network of power stations and transmission lines that deliver electricity across large spatial domains (e.g., country and continental scale) requires a variety of mathematical models, most of which are formulated as mathematical programs. Production Cost Models (PCM) are among the most common types of models found in the power systems literature: their task is to support the short-term operation of power grids [@garver1962threebin]. Specifically, PCMs schedule the production of generation fleets as well as the dispatch of electricity through the transmission lines over planning horizons that range from a few hours to a few days. 
 
-The framework is designed to be accessible to a wide range of users, especially those with only basic Python knowledge. To use `PowNet 2.0`, a user can supply data on power stations, transmission lines, and electricity demand as spreadsheets and run a script to handle the complex calculations and produce modeling outputs as either data frames or spreadsheet files. While a basic user does not need to touch the codebase, an advanced user can instead leverage the software’s flexible and modular design for complex tasks like modeling the interaction between water and power systems or exploring customized algorithms.
+`PowNet 2.0` is a Python-based PCM framework that simulates and analyzes the most cost-effective way of meeting electricity demand using various electricity sources (e.g., coal, natural gas, and renewables) such that sufficient electricity is generated at each hour while considering factors like fuel prices, availability of renewables, and regulatory requirements. The framework also allows us to study the integration between hydropower and power systems.
+
+The framework is designed to be accessible to a wide range of users, especially those with only basic Python knowledge. To use `PowNet 2.0`, a user can supply data on power stations, transmission lines, and electricity demand as spreadsheets, and run a script to handle the complex calculations and to produce modeling outputs as either data frames or spreadsheet files. While a basic user does not need to modify the codebase, an advanced user can instead leverage the software‚ flexible and modular design for more complex tasks, like modeling the interaction between water and power systems or exploring customized optimization algorithms.
 
 
 # Statement of need
 
[…]
