### 2025-05-19 Critical-Infrastructure-Systems-Lab/PowNet#22 @phumthep
@ijbd  @prakaa

Thank you for your thorough review and constructive feedback! Below, we address each of your comments:

Statement of Need

```
Seconding @prakaa's comments about which features to (re)prioritize, I think you have a great structure for comparing your software with the existing solutions. In thinking through your edits, you might consider identifying why [your] particular and future research lead(s) to pownet 2.0 filling gaps left by other solutions. This is partially done already by some references in your feature descriptions, but just a note as you edit.
```

We have revised the "Statement of Need" to more explicitly connect the development of PowNet 2.0 to the specific research gaps it aims to fill. The manuscript now clarifies that our research, particularly focusing on integrated water-power systems, stochastic uncertainty analysis, and the need for flexible modeling formulations, directly motivated the inclusion of features such as: (1) the flexibility to incorporate diverse mathematical formulations and solution algorithms, (2) the capacity to generate stochastic input data, and (3) the ability to integrate hydropower and power system models. These points are now more clearly presented as the core differentiators and motivators for PowNet 2.0's development, stemming from limitations encountered with existing tools in our specific research applications.

```
Line 36 and 40 "...the ability to efficiently handle large-scale systems with numerous nodes and edges..." and "...specifically designed to run multiple simulations of large-scale power systems..."
These are vague functional claims, the second more so than the first. If these are limitiations of existing packages, which of your features responds to these needs? Again, why did/does your research lead you to addressing these?
```

Following @prakaa's earlier suggestion, we have removed the phrase "...the ability to efficiently handle large-scale systems with numerous nodes and edges..." as we agree it did not sufficiently differentiate PowNet 2.0.
Regarding the statement "...specifically designed to run multiple simulations of large-scale power systems...", we have clarified this in the manuscript. This capability is directly linked to our "Stochastic generation of input variables" feature. Our research into the impacts of uncertainty in weather and climate on power systems necessitates running numerous simulations with varying inputs (e.g., load, renewable availability). PowNet 2.0 facilitates this by enabling the automated generation of these stochastic inputs and managing the ensemble runs. The manuscript now better reflects that this design choice was driven by the need to perform comprehensive uncertainty analyses, which often require evaluating a large number of scenarios.

```
Line 50 "Selecting a set of computationally efficient mathematical formulations is an on-going research Chen et al. (2022)"
Change to "research agenda Chen et al. (2022)"?
```

We have revised this sentence. It now reads: "Selecting a set of computationally efficient mathematical formulations is an ongoing research effort [@tejada2019unit; @chen2022security]." We have also included an additional relevant citation.

```
Line 63 "Reservoir simulation"
This is a very interesting feature, and all of the relevant studies and publications in the readthedocs appear to explore hydro and energy-water system coupling. Can you elaborate?
```

We agree that this is a key feature. The "Reservoir simulation" module is crucial for accurately modeling hydropower availability. It uses time series of reservoir inflow and dam design specifications (e.g., maximum storage, turbine capacity) as inputs. The model then calculates the hourly available hydropower based on a mass-balance equation and operational rules.

We have expanded the description in the manuscript to clarify that this integrated module offers two main advantages that were central to our research needs:

- Increased Detail and Accuracy: It avoids the oversimplification of using static capacity factors for hydropower.

- Reduced External Dependencies: It lessens the reliance on external, often complex, hydrologic models solely for deriving hydropower availability, making PowNet 2.0 more self-contained for power system studies that require dynamic hydropower representation.

```
Software Design, Implementation, and Validation

I don't see a discussion of validation, but former research publications and on-going research count! This could be a good place to list out those projects.
```

 We have revised the section to "Software design and implementation". We have also added a dedicated page on (PowNet Read the Docs)[https://pownet.readthedocs.io/en/latest/reference/publications.html] which lists studies that use PowNet. This allows us to showcase multiple studies that have used PowNet without being hindered by the word limit.

```
Line 67 "The software is specifically designed to support power-water systems coupling"
This is another indicator to me that power-water systems coupling should have more emphasis in the statement of need and possibly summary.
```

We have revised both the "Summary" and the "Statement of Need" to better emphasize the power-water systems coupling capabilities. In the Summary, we now explicitly state that a key capability of the framework is to "study the integration between hydropower and power systems". In the Statement of Need, the "ability to integrate hydropower and power system models" is highlighted as one of the three key features addressing gaps in existing PCM frameworks. The subsequent description of the "Reservoir simulation" and "Water-power system coupling" features further details this.

I hope our revision have addressed your comments! Please let us know if these are sufficient, so that we can close this issue.
