### 2025-05-19 #8035 @phumthep
@ijbd I have addressed your comments on the (issue tracker of PowNet's repo) as follow.
- [Manuscript revision](https://github.com/Critical-Infrastructure-Systems-Lab/PowNet/issues/22). Please let me know if there are other comments. Otherwise, I can close this issue.
- [Missing community guidelines](https://github.com/Critical-Infrastructure-Systems-Lab/PowNet/issues/21) I have added a Read the Docs page on contributing to PowNet 2.0 [here](https://pownet.readthedocs.io/en/latest/reference/contributing.html)
- [Broken links in README.md](https://github.com/Critical-Infrastructure-Systems-Lab/PowNet/issues/20) I have fixed all the broken links.
