### 2025-05-19 #8035 @phumthep
Hi @prakaa,

Please find our responses to your specific comments below:

```
In my opinion lines 8-12 in the summary (describing PCMs) could be condensed to improve readability and make the summary more concise. I think focusing on what PCMs do, as opposed to what they do not, would be helpful since dynamics and investment/planning models aren't described in the paper. A brief definition of "short-term" or the timeframes involved would add clarity too.
```

We have revised the summary to be more concise and focused. The description of Production Cost Models (PCMs) now concentrates on their core function: supporting the short-term operation of power grids by scheduling generation and dispatch. We have also clarified the "short-term" aspect by explicitly stating that the planning horizons typically "range from a few hours to a few days," as detailed later in the same paragraph.

```
There appears to be no space between the name of the software/package and the in-text citation for some of the software/packages in the Statement of need.
```

This has been fixed.

```
In the Mathematical Formulations section of the Statement of need, I think it would make more sense for these references to be in parentheses. You are not referring directly to the cited texts.
```

We have fixed the citations. They were incorrected formatted at first.

```
Examining the needs in lines 35-39, (1) and (4) are important but arguably handled relatively well by the software and packages you have listed in the preceding paragraph. Would it better highlight the contribution of this package to focus on (2) (solver flexibility & solution heuristics), (3) (module to generate synthetic time series to use as inputs) and on the water-energy focus (reservoir simulation - anything more?)? The Modelling Framework subheading could be incorporated into the summary.
```

We agree too. We have revised the "Statement of Need" to better emphasize the key differentiating features of PowNet 2.0. While flexibility in mathematical formulations remains an important aspect (previously part of your point 1), we have shifted the primary emphasis to:

- The ability to generate stochastic input data for uncertainty analysis (your point 3).

- The integration of hydropower and power system models, including detailed reservoir simulation and soft/hard coupling options (our water-energy focus, your point related to reservoir simulation).

- The potential to explore custom solution methods like heuristics (your point 2).

We have also removed the Modelling Framework subheading. We have also moved the discussion about Gurobipy to the "Software Design and Implementation" section, as it pertains more to the implementation choices rather than a unique need that PowNet 2.0 solely addresses.

```
Following on from the point above, do you think the listed features are the main points of differentiation from Sienna, PyPSA and PLEXOS?
```

Yes, we believe so. With the revisions emphasizing stochastic input generation, the detailed water-energy focus (including reservoir simulation and coupling methods), and solution methods, the "Statement of Need" now more clearly distinguish PowNet 2.0 from alternatives.

```
If the time series model in Figure 1 is linked to the ability to generate stochastic inputs as described in the Statement of need, I think it could be clearer in Figure 1 that this is the purpose of the "Time series model"
```

We have revised Figure 1. The box previously labeled "Time series model" is now more explicitly described "Stochastic Time Series Generation (from deterministic inputs)".

```
Figure 1 reference on line 70 should be fixed
```

Fixed as recommended.

I think these should address missing ticks in the "Software paper" section of the review checklist.
