### 2025-05-21 #8035 @phumthep
Hi @prakaa ,

To improve our documentation, we have totally revamped our Read the Docs website. These pages should sufficiently address your comments on documentation.

```
A brief mention of the target audience and/or comparison with similar packages could be handy for potential users
```

A section describing the target audience for PowNet 2.0 has been included in the "Introduction" page of our documentation: https://pownet.readthedocs.io/en/latest/getting_started/introduction.html.
This page also outlines key model functionalities for potential users to understand.

We opt to list our model functionalities instead of comparing with other similar packages to limit the scope of PowNet's documentation to itself. Also, each open-source package follows its own development pathway, catering to its community, making the monitor of their updates out-of-scope for us.

```
As a general comment, I think there could be more detail in the documentation. Some of this could be inferred from looking at the source code or the example input data for the dummy model, but I think it would be of general benefit to a user to have more detail. A non-exhaustive list of things that could be covered: format/requirements of input data (csvs/dataframes for costs, generators, nodes, lines, etc.), format of the output data and features mentioned in the JOSS paper (reservoir modelling, synthetic time series generation, the use of alternative formulations, integer rounding solution heuristic).
```

We have substantially expanded several sections:

- We now provide a dedicated section on "Preparing Input Files" which details the structure and format (primarily CSVs). This can be found at: https://pownet.readthedocs.io/en/latest/user_guide/input_files.html.

- We have modified and added new examples as Jupyter notebooks. For example, there is a detailed guide for a "Customized Modeling Workflow," which shows users how to interact with model outputs and plot results of interest: https://pownet.readthedocs.io/en/latest/examples/custom_workflow.html. Modeling examples also come with a description of the power system along with a figure of the system. These examples also cover specific features mentioned in the JOSS paper (reservoir modeling, synthetic time series generation, alternative formulations, solution heuristics).

- As for API documentation, we have fixed an issue with our Sphinx setup, so the API reference is now correctly generated for existing modules, classes, and methods: https://pownet.readthedocs.io/en/latest/api_ref/pownet.html. We have also revised and expanded docstrings throughout the codebase to improve clarity within the API reference.

```
The example did not work for me by just copying the code. The if statement that checks whether the output path exists and then creates it needs indentation. The input and output folder paths did not work on my machine (Ubuntu 24), could you use pathlib instead? I changed these and changed the solver to highs since I am no longer an academic with access to Gurobi, but it appears that the model can't be built without a Gurobi licence. Is this right (see screenshot below)
Image
```

We have added a specific instruction on where to download the input files in each notebook example. PowNet 2.0 was designed as a modeling framework, whereby the framework and the data are separate. Now, the Jupyter notebook examples emphasize that users need to define file paths relative to their local machine and project structure. In the examples, we use the `os` package to concatenate directories, which works across platforms similarly to the `pathlib` package. Users are free to use pathlib in their scripts

Regarding the issue with Gurobi license, we use Gurobipy to build instances of the optimization problem. This does not require a paid gurobi license. However, your case suggests that an expired license could cause this issue. Therefore, I would recommend a clean install of Gurobi and use only the trail license to run the example. Note that our examples are small enough to be solved with the trial license that comes with Gurobipy.

```
Similar to the documentation, the code example requires some additional guidance/explanation. First off, the user needs to clone the source code repo to access the dummy model directory (if they have pip installed or installed with a dependency management tool). Secondly, it's unclear what is being run - a summary of the scenario would be ideal, but at the very least a link and brief description of the inputs would help.
```

Our new Jupyter notebook examples have addressed these points. See our comments above.

```
In the API reference, some classes and methods are documented, but some are not. Is there some reason for this?
```

This issue is now resolved. We identified and fixed a problem with our Sphinx-autodoc configuration that was preventing some parts of the API from being rendered. The API reference at https://pownet.readthedocs.io/en/latest/api_ref/pownet.html now works correctly.

```
I couldn't find instructions for development - e.g. how to run tests.
I couldn't find contributing guidelines
```

We have now added a dedicated page on "Contributing to PowNet 2.0" in the documentation: https://pownet.readthedocs.io/en/latest/reference/contributing.html. This page includes guideline on contributing to the project, infor on running unit tests with the `unittest` package (specifically before submitting a pull request). Furthermore, we have also included a test coverage badge on our GitHub repository. While the current coverage is 38%, we have prioritized testing for core functionalities like constraint implementation. We are committed to progressively increasing test coverage for other parts of the codebase. Given the size and complexity of the codebase, this will be an ongoing effort.

Please let us know what remaining tasks are needed to complete all items on the checklist!
