### 2025-05-23 103c285
fix: typing of get_hourly_demand
