+++ src/pownet/core/visualizer.py
@@ -6,6 +6,7 @@ import contextily as cx
 import geopandas as gpd
 import matplotlib as mpl
 from matplotlib.colors import ListedColormap
+from matplotlib.lines import Line2D
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
@@ -62,7 +63,7 @@ class Visualizer:
             bbox_to_anchor=(0.5, -0.12),
         )
         ax.set_ylabel("Power (MW)")
-        ax.set_ylim(top=(demand[:total_timesteps].max() * 1.30).values[0])
+        ax.set_ylim(top=(demand[:total_timesteps].max() * 1.30))
 
         if output_folder is not None:
             figure_name = f"{self.model_id}_fuelmix.png"
@@ -395,4 +396,173 @@ class Visualizer:
         df_to_plot = contract_generation.sort_values(by=["value"], ascending=False)
         df_to_plot.plot(ax=ax, kind="bar", linewidth=2, legend=False)
         ax.set_ylabel("Generation (MWh)")
+        plt.show()
+
+    def plot_power_flow(self, 
+                        flow_variables: pd.DataFrame, 
+                        figsize_per_line: tuple = (10, 2), # Note: height is for plot area of one subplot
+                        fixed_legend_height_inches: float = 0.5) -> None:
+        """
+        Plots the power flow on transmission lines over time.
+
+        Each unique transmission line (node_a to node_b) gets its own subplot.
+        The y-axis label for each subplot is the line segment name.
+        Legend is placed at the top center of the figure, occupying a fixed absolute height.
+        Power flow is colored:
+        - Green: Positive flow
+        - Red: Negative flow
+        - Black: Zero flow
+
+        Args:
+            flow_variables (pd.DataFrame): DataFrame with simulation results. Expected columns:
+                                     'node_a', 'node_b', 'value', 'type' ('fwd' or 'bwd'), 'hour'.
+            figsize_per_line (tuple): Tuple specifying (width, height_for_each_subplot_plot_area).
+            fixed_legend_height_inches (float): Absolute height in inches for the legend area at the top.
+        """
+        lines = flow_variables[['node_a', 'node_b']].drop_duplicates().values.tolist()
+        num_lines = len(lines)
+
+        if num_lines == 0:
+            print("No transmission lines to plot.")
+            return
+
+        fig_width = figsize_per_line[0]
+        
+        # Calculate the total height needed for the plot areas of all subplots
+        plots_area_height_inches = figsize_per_line[1] * num_lines
+        if plots_area_height_inches < 0: # Ensure non-negative plot height
+            plots_area_height_inches = 0
+
+        # Calculate the total figure height, including the fixed space for the legend
+        total_figure_height_inches = plots_area_height_inches + fixed_legend_height_inches
+        
+        # Ensure total figure height is positive; if not, default to a minimum
+        if total_figure_height_inches <= 0:
+            total_figure_height_inches = max(fixed_legend_height_inches, 1.0) # Use legend height or 1 inch min
+
+        fig, axes = plt.subplots(num_lines, 1,
+                                 figsize=(fig_width, total_figure_height_inches), # Use the calculated total height
+                                 sharex=True, squeeze=False)
+        
+        # Calculate the fraction of the total figure height that the legend area will occupy
+        if total_figure_height_inches > 0: # Avoid division by zero
+            legend_top_margin_fraction = fixed_legend_height_inches / total_figure_height_inches
+        else: # Should be unreachable due to the check above
+            legend_top_margin_fraction = 0.5 # Fallback: 50% for legend if total height is still 0
+
[…]
