### 2025-05-23 757c6fe
fix: allow one demand node
