### 2025-05-23 Critical-Infrastructure-Systems-Lab/PowNet#21 @ijbd
Looks great!
