### 2025-05-24 Critical-Infrastructure-Systems-Lab/PowNet#22 @ijbd
@phumthep I like your edits to the paper! I attached some minor comments only for your consideration, but feel free to close this issue when you are ready.
