### 2025-05-24 #8035 @ijbd
@arfon I have completed my review. After their revisions, the paper is more focused on the features pownet 2 offers compared to other PCM models, and the documentation is much improved, including community/contributor guidelines. I left minor _optional_ follow-up comments on the paper for @phumthep to review. I will let them close Issue 22 when they are ready. Thank you!
