### 2025-05-24 67eba3b
fix: minor typing issue
