+++ src/pownet/core/model_builder.py
@@ -182,8 +182,8 @@ class ModelBuilder:
         return self.hydro_builder.phydro
 
     def update_daily_hydropower_capacity(
-        self, step_k: int, new_capacity: dict[(str, int), float]
-    ) -> None:
+        self, step_k: int, new_capacity: dict[tuple[str, int], float]
+    ) -> PowerSystemModel:
         """Update the daily hydro capacity in the model."""
         self.hydro_builder.update_daily_hydropower_capacity(step_k, new_capacity)
         self.model.update()
