### 2025-05-24 484d3c4
feat: add parse_lmp_dict
