### 2025-05-26 3f44260
docs: update JOSS paper
