### 2025-05-26 Critical-Infrastructure-Systems-Lab/PowNet#22 @phumthep
@ijbd Thanks for taking another look at our manuscript! Yes your comments make sense and we have modified our draft accordingly. with 3f44260.
