### 2025-05-28 712aa54
refactor: Create a coupler module
