+++ src/test_pownet/test_data_model/test_power_trade.py
@@ -1,0 +1,226 @@
+import unittest
+import os
+import tempfile
+import dataclasses
+
+
+from pownet.data_model.power_trade import PowerTradeParams
+
+# --- Test Data ---
+VALID_INTERTIE_POINTS = [("N1", "N2"), ("N3", "N4")]
+VALID_INTERTIE_CAPACITIES = {("N1", "N2"): 100.0, ("N3", "N4"): 200.0}
+VALID_CONNECTED_BA = [("BA1", "BA2"), ("BA3", "BA4")]
+VALID_TOTAL_TRANSFER_LIMITS = {("BA1", "BA2"): 150.0, ("BA3", "BA4"): 250.0}
+
+CSV_HEADER = "intertie_1_ba,intertie_1_node,intertie_2_ba,intertie_2_node,capacity_mw\n"
+VALID_CSV_DATA_ROW1 = "BA1,N1,BA2,N2,100.0\n"
+VALID_CSV_DATA_ROW2 = "BA1,N3,BA2,N4,50.5\n"
+VALID_CSV_DATA_ROW3 = "BA3,N5,BA4,N6,200.0\n"
+
+
+class TestPowerTradeParams(unittest.TestCase):
+    """Test suite for the PowerTradeParams dataclass and its methods."""
+
+    def _create_temp_csv_file(self, content: str) -> str:
+        """
+        Helper method to create a temporary CSV file with given content.
+        The file is registered for cleanup after the test.
+        Returns the path to the temporary file.
+        """
+        # Use mkstemp for a unique file name and then open it in text mode
+        fd, path = tempfile.mkstemp(suffix=".csv", text=True)
+        with os.fdopen(fd, "w") as tmp_file:
+            tmp_file.write(content)
+        self.addCleanup(
+            os.remove, path
+        )  # Ensures the file is deleted after the test method
+        return path
+
+    # --- __init__ and __post_init__ Tests ---
+
+    def test_successful_initialization(self):
+        """Test successful instantiation with valid parameters."""
+        params = PowerTradeParams(
+            intertie_points=VALID_INTERTIE_POINTS,
+            intertie_capacities=VALID_INTERTIE_CAPACITIES,
+            connected_ba=VALID_CONNECTED_BA,
+            total_transfer_limits=VALID_TOTAL_TRANSFER_LIMITS,
+        )
+        self.assertEqual(params.intertie_points, VALID_INTERTIE_POINTS)
+        self.assertEqual(params.intertie_capacities, VALID_INTERTIE_CAPACITIES)
+        self.assertEqual(params.connected_ba, VALID_CONNECTED_BA)
+        self.assertEqual(params.total_transfer_limits, VALID_TOTAL_TRANSFER_LIMITS)
+
+    def test_default_factory_initialization(self):
+        """Test instantiation with default factory values (empty lists/dicts)."""
+        params = PowerTradeParams()
+        self.assertEqual(params.intertie_points, [])
+        self.assertEqual(params.intertie_capacities, {})
+        self.assertEqual(params.connected_ba, [])
+        self.assertEqual(params.total_transfer_limits, {})
+
+    def test_post_init_negative_intertie_capacity(self):
+        """Test __post_init__ raises ValueError for negative intertie capacity."""
+        with self.assertRaisesRegex(ValueError, "cannot be negative"):
+            PowerTradeParams(intertie_capacities={("N1", "N2"): -100.0})
+
+    def test_post_init_self_loop_intertie(self):
+        """Test __post_init__ raises ValueError for self-loop in intertie_points."""
+        with self.assertRaisesRegex(ValueError, "Self-loop intertie found"):
+            PowerTradeParams(intertie_points=[("N1", "N1")])
+
+    def test_post_init_reverse_intertie(self):
[…]
