### 2025-06-01 #8035 @prakaa
Thanks @phumthep - great work, on addressing all the comments. The paper better highlights the package's unique features, and the new examples should be very useful to users.

The only things I'll raise is that it's a little unclear how nodes are defined from the inputs page on the docs. Does the first row of the demand csv define a name for the node, and then the other csvs refer to that same name?

Otherwise, my review is complete with the changes made to the paper and the docs. Happy for my side of the review to be closed out once the point above regarding nodes is clarified by the authors.
