### 2025-06-01 #8035 @phumthep
Thank you @prakaa! I 've added some notes in the user guide on this issue. This is addressed with commit 260bb9b.

When preparing PowNet's inputs for a system with only a single node, then the node is defined by `demand_export.csv`. However, if there are multiple nodes, then they are defined in `transmission.csv`.  There other CSV files would need to match the node names with these files accordingly.

>
> The only things I'll raise is that it's a little unclear how nodes are defined from the inputs page on the docs. Does the first row of the demand csv define a name for the node, and then the other csvs refer to that same name?

@arfon It appears that our reviewers have ticked all the boxes. Please let me know how best to proceed. Thank you!
