+++ src/pownet/data_model/__init__.py
@@ -1,7 +1,9 @@
 """This is the core module."""
 
 from .reservoir import ReservoirParams
+from .power_trade import PowerTradeParams
 
 __all__ = [
     "ReservoirParams",
+    "PowerTradeParams",
 ]
