+++ src/pownet/builder/system.py
@@ -38,6 +38,19 @@ class SystemBuilder(ComponentBuilder):
     ---------------------------
     - `spin_shortfall`: Spinning reserve shortfall (system-wide). Unit: MW.
 
+    Curtailment variables:
+    ---------------------------
+    - `pthermal_curtail`: Thermal unit curtailment. Unit: MW.
+    - `phydro_curtail`: Hydro unit curtailment. Unit: MW.
+    - `psolar_curtail`: Solar unit curtailment. Unit: MW.
+    - `pwind_curtail`: Wind unit curtailment. Unit: MW.
+    - `pimp_curtail`: Import unit curtailment. Unit: MW.
+
+    Trade variables:
+    ---------------------------
+    - `trade_fwd`: Forward trade on interties. Unit: MW.
+    - `trade_bwd`: Backward trade on interties. Unit: MW.
+
     Fixed objective terms
     ===========================
     - load shortfall penalty
@@ -77,6 +90,10 @@ class SystemBuilder(ComponentBuilder):
         self.pwind_curtail = gp.tupledict()
         self.pimp_curtail = gp.tupledict()
 
+        # Trade variables
+        self.trade_fwd = gp.tupledict()
+        self.trade_bwd = gp.tupledict()
+
         # --- Objective terms ---
         # Fixed objective terms
         self.load_shortfall_penalty_expr = gp.LinExpr()
@@ -109,6 +126,15 @@ class SystemBuilder(ComponentBuilder):
 
         self.c_daily_hydro_curtail_ess = gp.tupledict()
         self.c_weekly_hydro_curtail_ess = gp.tupledict()
+
+        # Trade constraints
+        # Note: Set the RHS of the opposite flow path to zero for mutual exclusivity
+        self.c_trade_fwd_ub = gp.tupledict() 
+        self.c_trade_bwd_ub = gp.tupledict()
+
+        # Trade LMPs
+        self.trade_lmps: dict[tuple[str, str], float] = {}
+        
 
     def add_variables(self, step_k: int) -> None:
 
@@ -201,6 +227,23 @@ class SystemBuilder(ComponentBuilder):
                 name="theta",
             )
 
+        # --- Trade variables ---
+        # UB set by constraints.
+        self.trade_fwd = self.model.addVars(
+            self.inputs.intertie_points,
+            self.timesteps,
+            lb=0,
+            vtype=gp.GRB.CONTINUOUS,
+            name="trade_fwd",
+        )
+        self.trade_bwd = self.model.addVars(
+            self.inputs.intertie_points,
+            self.timesteps,
+            lb=0,
+            vtype=gp.GRB.CONTINUOUS,
+            name="trade_bwd",
+        )
+
         #############################################
         # Variables with time-dependent upper bounds
         #############################################
@@ -221,7 +264,7 @@ class SystemBuilder(ComponentBuilder):
                 ]
                 for t in self.timesteps
                 for source, sink in self.inputs.edges
-            },
+            }, 
             vtype=gp.GRB.CONTINUOUS,
             name="flow_fwd",
         )
@@ -471,11 +514,13 @@ class SystemBuilder(ComponentBuilder):
             )
 
     def add_constraints(self, step_k: int, init_conds: dict, **kwargs) -> None:
-
         # Thermal-unit specific variables
         spin_vars = kwargs.get("spin_vars", None)
         vpowerbar_vars = kwargs.get("vpowerbar_vars", None)
         thermal_status_vars = kwargs.get("thermal_status_vars", None)
+
+        # Power trading
+        trade_limits = kwargs.get("trade_limits", {})
 
         # Power variables
         pthermal = kwargs.get("pthermal", None)
@@ -531,6 +576,8 @@ class SystemBuilder(ComponentBuilder):
[…]

[…]
