### 2025-06-03 #8035 @phumthep
Here are some additional tasks that I have also completed as well.

Double check authors and affiliations (including ORCIDs)
- I have checked that the authors and affiliations (including ORCIDs) are correctly described in the manuscript.

Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
- I have created a version 2.3 release on the GitHub repository [here](https://github.com/Critical-Infrastructure-Systems-Lab/PowNet/releases/tag/v2.3). I have also uploaded this version to the Python Package Index (PyPI) for users to install via `pip`. The link can be accessed [here](https://pypi.org/project/pownet/).

Archive the release on Zenodo/figshare/etc and post the DOI here.
- Bunnak, P., Eldardiry, H., Pavičević, M., Feng, J. Z., & Galelli, S. (2025). PowNet 2.0: A production cost modeling framework for large-scale power systems. In Journal of Open Source Software (2.3). Zenodo. https://doi.org/10.5281/zenodo.15580622

Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
- I confirm that the title and author’s information on Zenodo align with the JOSS manuscript.

Make sure that the license listed for the archive is the same as the software license.
- I confirm that the license listed for the archive is the same as the software license (MIT license).

Please let me know if I am missing anything.
