### 2025-06-05 cdc9d39
chore: Add acknowledgement to JOSS
Add HydroCHiPPs acknowledgement
