### 2025-06-05 c5cb043
chore: update acknowledgement in JOSS
