### 2025-06-11 92bf1e6
(fix): dispatch plot, allow for hydro_capacity_min to be empty (assign 0 if yes)
