### 2025-06-11 91394f4
fix(hydro): weekly_hydro_budget_min now optional
