+++ scripts/run_hydrochipps.py
@@ -1,3 +1,4 @@
+#%%
 import os
 from pownet.core import Simulator, DataProcessor, OutputProcessor
 
@@ -46,136 +47,137 @@ simulator.plot_fuelmix("area", output_folder)
 output_processor = OutputProcessor()
 hourly_demand = output_processor.get_hourly_demand(simulator.inputs.demand)
 
-## old code
-# #%%
-# import os
-# from datetime import datetime
-# from pownet.core import (
-#     DataProcessor,
-#     SystemInput,
-#     ModelBuilder,
-#     SystemRecord,
-#     OutputProcessor,
-#     Visualizer,
-# )
-# from pownet.data_utils import create_init_condition
-#
-# scen_path = ".\\model_library"
-# ##### User inputs #####
-# to_process_inputs = True
-# sim_horizon = 24*7
-# model_name = "cumberland"
-# steps_to_run = 3  # Default is None
-# do_plot = True
-# #######################
-#
-# if to_process_inputs:
-#     data_processor = DataProcessor(
-#         input_folder=scen_path, model_name=model_name, year=2016, frequency=50
-#     )
-#     data_processor.execute_data_pipeline()
-#
-# inputs = SystemInput(
-#     input_folder=scen_path,
-#     model_name=model_name,
-#     year=2016,
-#     sim_horizon=sim_horizon,
-#     spin_reserve_factor=0,
-#     load_shortfall_penalty_factor=1000,
-#     load_curtail_penalty_factor=1,
-#     spin_shortfall_penalty_factor=1000,
-#     line_capacity_factor=1,
-#     line_loss_factor=0
-# )
-# inputs.load_and_check_data()
-#
-#
-# model_builder = ModelBuilder(inputs)
-# record = SystemRecord(inputs)
-#
-# build_times = []
-# opt_times = []
-# objvals = []
-# init_conditions = create_init_condition(inputs.thermal_units) # init conditions should be empty dict if we dont have any thermal generators
-#
-# if steps_to_run is None:
-#     steps_to_run = 10  # 365 - (sim_horizon // 24 - 1)
-#
-# start_day = 1
-# for step_k in range(start_day, steps_to_run):
-#     start_time = datetime.now()
-#     if step_k == start_day:
-#         power_system_model = model_builder.build(
-#             step_k=step_k,
-#             init_conds=init_conditions,
-#         )
-#     else:
-#         power_system_model = model_builder.update(
-#             step_k=step_k,
-#             init_conds=init_conditions,
-#         )
-#     build_times.append((datetime.now() - start_time).total_seconds())
-#
-#     power_system_model.model.write(str(step_k) + '.lp')
-#     power_system_model.optimize(mipgap=0.001)
-#
-#     # Raise an error if the model is not feasible
-#     if not power_system_model.check_feasible():
-#         raise ValueError("Model is not feasible.")
-#
-#     objvals.append(power_system_model.get_objval())
-#     opt_times.append(power_system_model.get_runtime())
-#
-#     record.keep(
-#         runtime=power_system_model.get_runtime(),
-#         objval=power_system_model.get_objval(),
-#         solution=power_system_model.get_solution(),
-#         lmp=power_system_model.solve_for_lmp(),
-#         step_k=step_k,
-#     )
-#     init_conditions = record.get_init_conds()
-#
-# import os
+# old code
+#%%
+import os
+from datetime import datetime
+from pownet.core import (
+    DataProcessor,
+    Simulator,
+    ModelBuilder,
+    SystemRecord,
+    OutputProcessor,
+    Visualizer,
+)
+from pownet import SystemInput
+from pownet.data_utils import create_init_condition
+
+scen_path = ".\\model_library"
+##### User inputs #####
+to_process_inputs = True
[…]
