### 2025-06-18 8cbbb9a
feat(hydro): lower bound for power output of hydro units
