### 2025-06-23 5c291e9
feat(loops): execute in a rolling horizon loop, functions to update budgets and prices
