+++ scripts/run_hydrochipps.py
@@ -1,55 +1,29 @@
 #%%
+"""
+Hydropower Scheduling Using Pownet (HydroCHIPPs Project)
+
+This script performs short-term hydropower scheduling for the Cumberland River Basin
+using the PoWnet modeling framework in rolling-horizon mode.
+
+Hydropower Units (underscore used instead of spaces for compatibility with HiGHS solver):
+    - Barkley
+    - Center_Hill
+    - Cheatham
+    - Cordell_Hull
+    - Dale_Hollow
+    - J_Percy_Priest
+    - Old_Hickory
+    - Wolf_Creek
+
+Authors:
+    - M. Pavičević, Ph.D. (Argonne National Laboratory)
+    - P. Bunnak (Cornell University)
+
+Project:
+    - HydroCHIPPs (DoE-sponsored HydroWIRES project)
+"""
 import os
-from pownet.core import Simulator, DataProcessor, OutputProcessor
-
-##### User inputs #####
-scen_path = ".\\model_library"
-results_path = ".\\results"
-to_process_inputs = True
-sim_horizon = 24*7
-model_name = "cumberland"
-steps_to_run = 3  # Default is None
-do_plot = True
-
-
-if to_process_inputs:
-    data_processor = DataProcessor(
-        input_folder=scen_path, model_name=model_name, year=2016, frequency=50
-    )
-    data_processor.execute_data_pipeline()
-
-simulator = Simulator(
-    input_folder=scen_path,
-    model_name=model_name,
-    model_year=2016,
-    spin_reserve_factor=0,
-    load_shortfall_penalty_factor=1000,
-    load_curtail_penalty_factor=1,
-    spin_shortfall_penalty_factor=1000,
-    line_capacity_factor=1,
-    line_loss_factor=0
-)
-
-simulator.run(
-    sim_horizon=sim_horizon,
-    steps_to_run=steps_to_run,
-    # solver='gurobi',
-)
-
-# Write the simulation results
-output_folder = results_path + '\\' + model_name
-if not os.path.exists(output_folder):
-    os.makedirs(output_folder)
-simulator.write_results(output_folder)
-
-simulator.plot_fuelmix("area", output_folder)
-
-output_processor = OutputProcessor()
-hourly_demand = output_processor.get_hourly_demand(simulator.inputs.demand)
-
-# old code
-#%%
-import os
+import numpy as np
 from datetime import datetime
 from pownet.core import (
     DataProcessor,
@@ -62,21 +36,98 @@ from pownet.core import (
 from pownet import SystemInput
 from pownet.data_utils import create_init_condition
 
-scen_path = ".\\model_library"
-##### User inputs #####
-to_process_inputs = True
-sim_horizon = 24*7
-model_name = "cumberland"
-steps_to_run = 3  # Default is None
-do_plot = True
-#######################
+# === File Path and Simulation Controls ===
+scen_path = "./model_library"      # Path to input data
+model_name = "cumberland"          # Name of the model (scenario folder)
+sim_horizon = 24 * 7               # Simulation horizon in hours (e.g., 7 days)
+steps_to_run = 2                   # How many daily steps to simulate per iteration
+end_day = 7                        # How many days to loop through in total
+to_process_inputs = True           # If True, runs the data processing pipeline
+do_plot = True                     # Currently unused, placeholder for plotting
 
+
+#%%###################################################
+## Functions to update the weekly budgets and prices
+######################################################
+def update_weekly_hydro_capacity(df, index, updated_values):
+    """
+    Update weekly hydro generation budgets in the DataFrame.
+
+    Parameters:
+    - df (pd.DataFrame): DataFrame containing weekly hydro capacities.
+    - index (int): Row index to update (typically the day number).
+    - updated_values (dict): Unit-wise capacity values for the specified index.
[…]
