### 2025-06-28 002f81d
fix: missing hydro_node error message
