### 2025-07-14 40462e2
fix: area plot issue
