### 2025-07-14 c20a023
fix: fix stability limit and remove user susceptance
