### 2025-07-14 #8035 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
