### 2025-07-15 750e0ed
chore: fix joss references
