+++ joss_submission/paper.bib
@@ -6,6 +6,7 @@
   number={3},
   pages={730--734},
   year={1962},
+  doi={10.1109/aieepas.1962.4501405},
   publisher={IEEE}
 }
 @article{horsch2018linear,
@@ -15,6 +16,7 @@
   volume={158},
   pages={126--135},
   year={2018},
+  doi={10.1016/j.epsr.2017.12.034},
   publisher={Elsevier}
 }
 @article{knueven2020mixed,
@@ -25,6 +27,7 @@
   number={4},
   pages={857--876},
   year={2020},
+  doi={10.1287/ijoc.2019.0944},
   publisher={INFORMS}
 }
 @article{koh2022hard,
@@ -34,6 +37,7 @@
   volume={321},
   pages={119386},
   year={2022},
+  doi={10.1016/j.apenergy.2022.119386},
   publisher={Elsevier}
 }
 @article{galelli2022opportunities,
@@ -44,6 +48,7 @@
   number={12},
   pages={1058--1069},
   year={2022},
+  doi={10.1038/s41893-022-00971-z},
   publisher={Nature Publishing Group UK London}
 }
 @article{chowdhury2020expected,
@@ -54,6 +59,7 @@
   number={10},
   pages={05020019},
   year={2020},
+  doi={10.1061/(ASCE)WR.1943-5452.0001279},
   publisher={American Society of Civil Engineers}
 }
 @article{chowdhury2021greater,
@@ -64,6 +70,7 @@
   number={3},
   pages={e2020EF001814},
   year={2021},
+  doi={10.1029/2020EF001814},
   publisher={Wiley Online Library}
 }
 @article{arnold2024floating,
@@ -72,6 +79,7 @@
   journal={Nature Energy},
   pages={1--10},
   year={2024},
+  doi={10.1038/s41560-024-01510-0},
   publisher={Nature Publishing Group UK London}
 }
 
@@ -89,6 +97,7 @@
   volume={102},
   pages={234--248},
   year={2019},
+  doi={10.1016/j.rser.2018.11.020},
   publisher={Elsevier}
 }
 @article{hofmann2023linopy,
@@ -98,6 +107,7 @@
   volume={8},
   number={84},
   pages={4823},
+  doi={10.21105/joss.04823},
   year={2023}
 }
 @article{chen2016improving,
@@ -108,6 +118,7 @@
   number={6},
   pages={4732--4743},
   year={2016},
+  doi={10.1109/TPWRS.2016.2530811},
   publisher={IEEE}
 }
 @article{su2020open,
@@ -117,6 +128,7 @@
   volume={126},
   pages={104667},
   year={2020},
+  doi={10.1016/j.envsoft.2020.104667},
   publisher={Elsevier}
 }
 @article{hoffmann2024review,
@@ -125,6 +137,7 @@
   journal={Advances in Applied Energy},
   pages={100190},
   year={2024},
+  doi={10.1016/j.adapen.2024.100190},
   publisher={Elsevier}
 }
 @article{tejada2019unit,
@@ -135,6 +148,7 @@
   number={4},
   pages={2926--2936},
   year={2019},
+  doi={10.1109/TPWRS.2019.2962024},
   publisher={IEEE}
 }
 @article{oberle2019open,
@@ -144,6 +158,7 @@
   volume={26},
   pages={100396},
   year={2019},
+  doi={10.1016/j.esr.2019.100396},
   publisher={Elsevier}
 }
 @article{oikonomou2022core,
@@ -153,6 +168,7 @@
   volume={238},
   pages={122049},
   year={2022},
+  doi={10.1016/j.energy.2021.122049},
   publisher={Elsevier}
 }
 @article{chen2022security,
@@ -163,6 +179,7 @@
   number={5},
   pages={4668--4681},
   year={2022},
[…]
