### 2025-07-15 #8035 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
